==> analysis/05_proteinPanel_StromalSubclustering.Rmd <==
---
title: "05_proteinPanel_StromalSubclustering"
author: "SandraTietscher"
date: "2020-10-23"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This script performs subclustering of stromal cells for the Protein panel. It further subclusters T/NK cells.
Finally, it merges cell type labels back into the original object, resulting in the annotated full object.

# Load libraries and data

```{r read-libraries-and-data, message=FALSE, results="hide"}
# Load libraries
library(ggplot2)
library(SingleCellExperiment)
library(scater)
library(tidyverse)
library(dittoSeq)
library(viridis)
library(cowplot)
library(ggpubr)
library(scran)
library(plyr)
library(patchwork)

# Read SingleCellExperiment object (with stromal cells only)
sce_stromal <- readRDS(paste0(getwd(), "/output/ProteinPanel/sce_ProteinPanel_stroma.rds"))
```

# Stromal cell subclustering

## Calculate UMAP on all stromal cells

```{r stromal-umap}
markers_exclude_umap <- c("DNA1", "DNA2", "HH3", "LAG-3", "IDO1", "PD1", "Tim-3", "MMP9", "Granzyme_B", "CD274", "TCF7", "ICOS", "CAIX", "GITR", "Arginase", "CD40", "iNOS", "Cleaved_CP", "Ki-67", "E-cadherin", "panCK")
markers_include_umap <- rownames(assay(sce_stromal, "counts"))[! (rownames(assay(sce_stromal, "counts")) %in% markers_exclude_umap)]

sce_stromal <- runUMAP(sce_stromal, exprs_values = "exprs", 
                     subset_row = rownames(assay(sce_stromal, "scaled")) %in% markers_include_umap)
```

### Visualize features on UMAP

```{r stromal-umap-features, fig.width=8, fig.height=8}

dittoDimPlot(sce_stromal, var = "sample", reduction.use = "UMAP", size = 0.5, 
                               color.panel = metadata(sce_stromal)$colour_vectors$sample, opacity = 0.2) 

dittoDimPlot(sce_stromal, var = "sample", reduction.use = "UMAP", size = 0.5, 
                               color.panel = metadata(sce_stromal)$colour_vectors$sample, split.by = "sample", opacity = 0.2) 

dittoDimPlot(sce_stromal, var = "TIG", reduction.use = "UMAP", size = 0.5, 
                               color.panel = metadata(sce_stromal)$colour_vectors$TIG)

dittoDimPlot(sce_stromal, var = "TLS", reduction.use = "UMAP", size = 0.5, 
                               color.panel = metadata(sce_stromal)$colour_vectors$TLS)
```

### Visualize marker expression on UMAP

```{r stromal-umap-markerexpression, fig.width=35, fig.height=25, message=FALSE, warning=FALSE}

# Plot expression of all markers
p.list <- list()
for(i in rownames(sce_stromal)){
  p.list[[i]] <- plotUMAP(sce_stromal, colour_by = i, by_exprs_values = "scaled", 
                          point_size = 0.5, point_alpha = 1, shape_by = "panel_type")  + scale_colour_gradient2(name = i, low = "#2166ac", mid = "white", high = "#b2182b", limits = c(-3, 3), na.value = "#b2182b")
}
plot_grid(plotlist = p.list, ncol = 7)
```

## Run graph-based clustering for stromal subset (Phenograph)

Display clusters on UMAP.

```{r stromal-clustering-umap, fig.width=8, fig.height=8}
cluster_counts_stromal <- subset(assay(sce_stromal, "normalized"), rownames(assay(sce_stromal, "normalized")) %in% markers_include_umap)

g <- buildSNNGraph(cluster_counts_stromal, k=7)
clust <- igraph::cluster_louvain(g)$membership
table(clust)
write.csv(table(clust), paste0(getwd(), "/output/ProteinPanel/stromal_cluster_cellnumbers.csv"), row.names = FALSE)

#add cluster information to sce object
sce_stromal$stromal_cluster <- as.factor(clust)

#plot on umap
dittoDimPlot(sce_stromal, var = "stromal_cluster", reduction.use = "UMAP", size = 0.2, do.label = TRUE, opacity = 0.2)
dittoDimPlot(sce_stromal, var = "stromal_cluster", reduction.use = "UMAP", size = 0.2, split.by = "stromal_cluster", opacity = 0.2)
```

### Plot mean expression for each cluster

```{r stromal-clusters-expression-heatmap, fig.width=5, fig.height=8}
clustermean <- aggregateAcrossCells(sce_stromal, ids = DataFrame(cluster = sce_stromal$stromal_cluster), average = TRUE)
colnames(clustermean) <- clustermean$cluster

# Center and scale raw counts
assay(clustermean, "scaled") <- t(scale(t(assay(clustermean, "counts"))))

# Normalize raw counts
norm_counts <- t(apply(assay(clustermean, "counts"), 1, function(x)(x-min(x))/(max(x)-min(x))))
assay(clustermean, "normalized", withDimnames = FALSE) <- norm_counts

## Plot only relevant markers
exclude_plot <- c("DNA1", "DNA2", "HH3")
include_plot <- rownames(assay(sce_stromal, "counts"))[! (rownames(assay(sce_stromal, "counts")) %in% exclude_plot)]

## Plot
# Z-scaled
dittoHeatmap(clustermean[include_plot,], assay = "scaled", scale = "none",
            heatmap.colors = magma(100), 
            cluster_cols = TRUE, show_colnames = TRUE,
            breaks = seq(-3, 3, length.out = 101))

# Normalized (0-1)
dittoHeatmap(clustermean[include_plot,], assay = "normalized", scale = "none",
            heatmap.colors = viridis(100), 
            cluster_cols = TRUE, show_colnames = TRUE)
```

### Assign cell type labels

Manual cell type assignment based on canonical marker expression.

```{r stromal-labels, fig.width=8, fig.height=8}
labels <- read.csv(paste0(getwd(), "/data/ProteinPanel_stromal_clusters.csv"))

label_list <- labels$class[match(sce_stromal$stromal_cluster, labels$cluster)]
sce_stromal$celltype <- label_list

# Plot on UMAP
dittoDimPlot(sce_stromal, var = "celltype", reduction.use = "UMAP", size = 0.5, opacity = 0.5, do.label = TRUE)
```

```{r stromal-proportions, fig.width=8, fig.height=6}
#Plot subtype proportions per sample
tab <- as.data.frame(prop.table(table(sce_stromal$sample, sce_stromal$celltype), margin = 1))
colnames(tab) <- c("sample", "celltype", "proportion")

ggplot(tab, aes(sample, y = proportion, fill=celltype)) +
  geom_bar(stat="identity")+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))
```

# Combine all celltype labels into full SCE object

```{r combine-celltype-labels}
# read in objects
sce_Protein <- readRDS(paste0(getwd(), "/output/ProteinPanel/sce_ProteinPanel_TS_assigned.rds"))
sce_Protein_sub <- readRDS(paste0(getwd(), "/output/ProteinPanel/sce_ProteinPanel_sub_UMAP.rds"))


## Prepare dataframe with cell IDs and celltype labels ##

# Tumor cells
tumor_ids <- as.data.frame(colnames(sce_Protein)[which(sce_Protein$cell_class == "tumor")])
colnames(tumor_ids) <- "cellID"
tumor_ids$celltype <- "tumor"

# Stromal cells without T/NK
stromal_ids <- data.frame(cellID = colnames(sce_stromal), celltype = sce_stromal$celltype)

# Combine
ids_table <- rbind(tumor_ids, stromal_ids)

# Add to full object
cellids_ordered <- as.data.frame(colnames(sce_Protein))
colnames(cellids_ordered)<- "cellID"
cellids_merged <- join(cellids_ordered, ids_table, by = "cellID")
sce_Protein$celltype <- cellids_merged$celltype

```

## Plot celltypes on UMAP

Use the randomly subset SCE object (120.000 cells).

```{r plot-celltype-umap, fig.width=8, fig.height=8}
# Assign cluster numbers to the right cells
sub_cellnames <- colnames(sce_Protein_sub)
celltype_sub <- sce_Protein[,sub_cellnames]$celltype
sce_Protein_sub$celltype <- celltype_sub

#plot on umap
dittoDimPlot(sce_Protein_sub, var = "celltype", reduction.use = "UMAP", size = 0.5, do.label = TRUE)
```

## Plot expression heatmap with celltype labels

```{r expression_heatmap_celltype_labels}

clustermean <- aggregateAcrossCells(sce_stromal, ids = DataFrame(cluster = sce_stromal$stromal_cluster), average = TRUE)
colnames(clustermean) <- clustermean$cluster

# Normalize raw counts
norm_counts <- t(apply(assay(clustermean, "counts"), 1, function(x)(x-min(x))/(max(x)-min(x))))
assay(clustermean, "normalized", withDimnames = FALSE) <- norm_counts

dittoHeatmap(clustermean[markers_include_umap,], assay = "normalized", scale = "none",
            heatmap.colors = viridis(100),
            cluster_cols = TRUE, show_colnames = TRUE, annot.by = "celltype", annot.colors =
              c("#7F00FF", "#FF3399", "#009900", "#3333FF", "#FF0000", "#CC99FF", "#FFFF33", "#FF9933", "#C0C0C0", "#00FF00"))
```


## Plot celltype proportions per sample and per image

Include TLS and pre-TLS images.

```{r celltype-proportion-sample, fig.width=6, fig.height=4}
celltype_colors = c(tumor = "#00FFFF",
                          fibroblast = "#3333FF",
                          endothelial = "#009900",
                          T_NK = "#00FF00",
                          pDC = "#FFFF33",
                          plasma_cell = "#FF9933",
                          myeloid = "#FF0000",
                          aDC = "#7F00FF",
                          B_cell = "#FF3399",
                          neutrophil = "#CC99FF",
                          stromal_undefined = "#C0C0C0"
                          )

# Plot cell type frequencies per sample
tab <- as.data.frame(prop.table(table(sce_Protein$sample, sce_Protein$celltype), margin = 1))
colnames(tab) <- c("sample", "celltype", "proportion")
tab$celltype <- factor(tab$celltype, levels = rev(names(celltype_colors)))

ggplot(tab, aes(sample, y = proportion, fill=celltype)) +
  geom_bar(stat="identity")+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))+
  scale_fill_manual(values = celltype_colors)
```

```{r celltype-proportion-image, fig.width=8, fig.height=3}
# Plot cell type frequencies per image
tab <- as.data.frame(prop.table(table(sce_Protein$ImageNumber, sce_Protein$celltype), margin = 1))
colnames(tab) <- c("image", "celltype", "proportion")

# Order images by tumor cell content
tumor_tab <- filter(tab, celltype == "tumor")
order <- tumor_tab[order(tumor_tab$proportion),]$image
tab$image <- factor(tab$image, levels = order)
tab$celltype <- factor(tab$celltype, levels = rev(names(celltype_colors)))

p_proportion <- ggplot(tab, aes(image, y = proportion, fill=celltype)) +
  geom_bar(stat="identity")+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))+
  scale_fill_manual(values = celltype_colors)

# Plot image metadata
ROI_info <- read.csv("data/ROI_info_HIER_TLSrevised.csv")
ROI_info$RNApanel_ImageNumber <- factor(ROI_info$RNApanel_ImageNumber, levels = order)
ROI_info$sample <- factor(ROI_info$sample, levels = levels(sce_stromal$sample))

p_sample <- ggplot(ROI_info, aes(x = RNApanel_ImageNumber, y = "sample", fill = sample))+
  geom_tile()+
  scale_fill_manual(values = metadata(sce_stromal)$colour_vectors$sample)+
  theme(panel.background = element_blank(), axis.text.x = element_blank(), axis.title = element_blank(), axis.ticks = element_blank())

p_TLS <- ggplot(ROI_info, aes(x = RNApanel_ImageNumber, y = "TLS", fill = TLS))+
  geom_tile()+
  scale_fill_manual(values = metadata(sce_stromal)$colour_vectors$TLS)+
  theme(panel.background = element_blank(), axis.text.x = element_blank(), axis.title = element_blank(), axis.ticks = element_blank())

p_TIG <- ggplot(ROI_info, aes(x = RNApanel_ImageNumber, y = "TIG", fill = TIG))+
  geom_tile()+
  scale_fill_manual(values = metadata(sce_stromal)$colour_vectors$TIG)+
  theme(panel.background = element_blank(), axis.text.x = element_blank(), axis.title = element_blank(), axis.ticks = element_blank())

p_TLS/p_TIG/p_sample/p_proportion + plot_layout(heights = c(1,1,1,20), guides = "collect")
```

## Compare celltype proportions per TIG

Exclude TLS and pre-TLS.

```{r celltype-proportions-TIG-boxplot, fig.width=10, fig.height=8}
sce_TLSex <- sce_Protein[,which(sce_Protein$TLS == "no")]
tab <- as.data.frame(prop.table(table(sce_TLSex$sample, sce_TLSex$celltype), margin = 1))
colnames(tab) <- c("sample", "celltype", "proportion")
TIG2 <- c("TBB111", "TBB129", "TBB165", "TBB171", "TBB184", "TBB338")
tab$TIG <- ifelse(tab$sample %in% TIG2, "TIG2", "TIG3")

ggplot(tab, aes(x= TIG, y = proportion, color=TIG))+
  geom_boxplot()+
  geom_point()+
  facet_wrap(~celltype, scales="fixed", ncol=6, strip.position="top")+
  theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Of total [%]")+
  stat_compare_means(method = "wilcox.test", aes(label=..p.format..), label.x.npc = "middle", label.y.npc = 0.9)
```


# Save object

```{r save-objects}
saveRDS(sce_Protein, "output/ProteinPanel/sce_ProteinPanel_celltypes.rds")
```


# Compare cell type proportions from CyTOF, RNAseq and IMC

Celltypes that were found in all three methods:  
-epithelial  
-endothelial  
-T_NK  
-myeloid (incl. dendritic cells)  
-B cell  
-plasma cell  
-fibroblast  
  
"Other" includes granulocytes, neutrophils, and "stromal_undefined" cell types.  
  
For IMC data: include TLS and pre-TLS images.  

## Prepare tables

```{r prepare-celltype-comparison-methods}

## Adapt IMC celltypes ##
sce_IMC <- sce_Protein
#sce_IMC <- sce_IMC[,which(sce_IMC$TLS == "no")]
sce_IMC$celltype <- factor(sce_IMC$celltype, levels=c("aDC", "B_cell", "endothelial", "fibroblast", "myeloid", "neutrophil", "pDC", "plasma_cell", "stromal_undefined", "T_NK", "tumor", "other"))
sce_IMC[,which(sce_IMC$celltype == "pDC")]$celltype <- "myeloid"
sce_IMC[,which(sce_IMC$celltype == "aDC")]$celltype <- "myeloid"
sce_IMC[,which(sce_IMC$celltype == "neutrophil")]$celltype <- "other"
sce_IMC[,which(sce_IMC$celltype == "stromal_undefined")]$celltype <- "other"
sce_IMC$celltype <- droplevels(sce_IMC$celltype)

IMC_perc <- as.data.frame(prop.table(x=table(sce_IMC$sample, sce_IMC$celltype), margin=1))
colnames(IMC_perc) <- c("sample", "celltype", "proportion")
IMC_perc$method <- "IMC"
IMC_perc$celltype <- gsub("tumor", "epithelial", IMC_perc$celltype)
# Remove sample not present in CyTOF
IMC_perc <- filter(IMC_perc, sample != "TBB338")

## Read in and adapt CyTOF celltypes ##
cytof_perc <- read.csv(paste0(getwd(), "/data/cytof_celltype_prop.csv"), row.names = 1)
cytof_perc["other",] <- cytof_perc["granulocyte",] + cytof_perc["other",]
cytof_perc["myeloid", ] <- cytof_perc["myeloid", ] + cytof_perc["dendritic cell", ]
cytof_perc <- cytof_perc[!(rownames(cytof_perc) %in% c("dendritic cell", "granulocyte")),]
rownames(cytof_perc) <- c("endothelial", "epithelial", "fibroblast", "B_cell", "plasma_cell", "myeloid", "other", "T_NK")
cytof_perc$celltype <- rownames(cytof_perc)
cytof_perc <- pivot_longer(cytof_perc, 1:12, names_to = "sample", values_to = "proportion")
cytof_perc$method <- "CyTOF"
cytof_perc <- cytof_perc[,c(2,1,3,4)]
# Remove sample not present in IMC
cytof_perc <- filter(cytof_perc, sample != "TBB011")


## Read in and adapt RNA-seq celltypes ##
RNAseq_perc <- read.csv(paste0(getwd(), "/data/RNAseq_celltype_prop.csv"))
# Remove samples not present in CyTOF or IMC
RNAseq_perc <- filter(RNAseq_perc, !(sample == "TBB338" | sample == "TBB011" | sample == "TBB330"))
RNAseq_perc <- pivot_longer(RNAseq_perc, 2:9, names_to = "celltype", values_to = "proportion")
RNAseq_perc$method <- "RNAseq"

# Combine
celltype_perc <- rbind(IMC_perc, cytof_perc, RNAseq_perc)
celltype_perc$method <- factor(celltype_perc$method, levels = c("IMC", "CyTOF", "RNAseq"))
```

## Plot

```{r celltype-comparison-barplot, fig.width=8, fig.height=5}
# Plot
ggplot(celltype_perc, aes(method, y = proportion, fill=celltype)) +
  geom_bar(stat="identity")+
  facet_wrap(~sample, ncol=6)+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))
```

```{r celltype-comparison-paired, fig.width=8, fig.height=5}
# Plot
ggplot(celltype_perc, aes(x = method, y = proportion)) +
  geom_line(aes(group = sample, color = sample))+
  scale_color_manual(values = sce_IMC@metadata$colour_vectors$sample)+
  facet_wrap(~celltype, scales = "free", ncol = 4)+
  theme(axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 30),
        panel.background = element_blank(),
        strip.background = element_blank(),
        legend.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1))

# IMC-RNAseq comparison
ggplot(celltype_perc[which(celltype_perc$method != "CyTOF"),], aes(x = method, y = proportion)) +
  geom_line(aes(group = sample, color = sample))+
  facet_wrap(~celltype, scales = "free", ncol = 4)+
  theme(axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 30),
        panel.background = element_blank(),
        strip.background = element_blank(),
        legend.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1))

# IMC-CyTOF comparison
ggplot(celltype_perc[which(celltype_perc$method != "RNAseq"),], aes(x = method, y = proportion)) +
  geom_line(aes(group = sample, color = sample))+
  facet_wrap(~celltype, scales = "free", ncol = 4)+
  theme(axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 30),
        panel.background = element_blank(),
        strip.background = element_blank(),
        legend.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1))
```

# Compare cell type proportions of the two panels

```{r celltype-comparison-panels, fig.width=15, fig.height=10}
sce_RNA <- readRDS("output/RNApanel/sce_RNApanel_celltypes.rds")

sce_RNA[,which(sce_RNA$celltype %in% c("stromal_SMA+", "stromal_undefined"))]$celltype <- "stromal"
sce_RNA[,which(sce_RNA$celltype == "Bcell")]$celltype <- "B_cell"
sce_RNA[,which(sce_RNA$celltype %in% c("NK", "T_CD4", "T_CD8"))]$celltype <- "T_NK"

sce_Protein[,which(sce_Protein$celltype %in% c("fibroblast", "stromal_undefined", "plasma_cell", "pDC"))]$celltype <- "stromal"

tab_protein <- as.data.frame(prop.table(table(sce_Protein$ImageNumber, sce_Protein$celltype), margin = 1))
tab_protein$panel <- "protein_panel"
tab_RNA <- as.data.frame(prop.table(table(sce_RNA$ImageNumber, sce_RNA$celltype), margin = 1))
tab_RNA$panel <- "RNA_panel"

tab_comb <- rbind(tab_protein, tab_RNA)
colnames(tab_comb) <- c("ImageNumber", "celltype", "proportion", "panel")

# Plot
ggplot(tab_comb, aes(panel, y = proportion, fill=celltype)) +
  geom_bar(stat="identity")+
  facet_wrap(~ImageNumber, ncol = 20)+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))
```

```{r celltype-comparison-panels-samplebased, fig.width=10, fig.height=7}
tab_protein <- as.data.frame(prop.table(table(sce_Protein$sample, sce_Protein$celltype), margin = 1))
tab_protein$panel <- "protein_panel"
tab_RNA <- as.data.frame(prop.table(table(sce_RNA$sample, sce_RNA$celltype), margin = 1))
tab_RNA$panel <- "RNA_panel"

tab_comb <- rbind(tab_protein, tab_RNA)
colnames(tab_comb) <- c("sample", "celltype", "proportion", "panel")

# Plot
ggplot(tab_comb, aes(panel, y = proportion, fill=celltype)) +
  geom_bar(stat="identity")+
  facet_wrap(~sample, ncol = 20)+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))
```

# MMP9 expression of myeloid cells

```{r myeloid_MMP9, fig.width=3, fig.height=3.5}
# sce_myeloid <- sce_protein[,which(sce_protein$celltype == "myeloid")]
# 
# counts <- data.frame(MMP9 = assay(sce_myeloid, "counts")["MMP9",])
# tab <- cbind(counts, colData(sce_myeloid))
# MMP9.means <- tab %>% group_by(sample) %>% dplyr::summarise(mean = mean(MMP9))
# MMP9.means$TIG <- c(rep("TIG2", 6), rep("TIG3", 6))
# 
# ggplot(MMP9.means, aes(TIG, mean, color = TIG))+
#     geom_boxplot(aes(color=TIG))+
#     geom_point(position=position_dodge(width=0.75), aes(group=TIG))+
#     stat_compare_means(method = "wilcox.test", aes(label=..p.format..), vjust = 1, label.x.npc = 0.5)+
#     theme(axis.line.x = element_line(colour = "black", size = 0.25),
#         axis.line.y = element_line(colour = "black", size = 0.25),
#         panel.grid.major = element_blank(),
#         panel.grid.minor = element_blank(),
#         panel.border = element_rect(colour = "black", fill=NA, size=1),
#         panel.background = element_blank(),
#         strip.background = element_blank())+
#   ylab("Mean counts")+
#   ggtitle("MMP9 expression on myeloid cells (patient means)")
```
