==> analysis/07_proteinPanel_TNK_analysis.Rmd <==
---
title: "07_proteinPanel_TNK_analysis"
author: "SandraTietscher"
date: "2020-10-27"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
```

This script performs subclustering of T cells and NK cells for the Protein panel and performs further downstream analysis

# Load libraries and data

```{r read-libraries-and-data, message=FALSE, results="hide"}
# Load libraries
library(ggplot2)
library(SingleCellExperiment)
library(scater)
library(tidyverse)
library(dittoSeq)
library(viridis)
library(cowplot)
library(ggpubr)
library(scran)
library(CValternatives)

# Read SingleCellExperiment object 
sce_protein <- readRDS("output/ProteinPanel/sce_ProteinPanel_celltypes.rds")

# Subset T/NK cells only
sce_TNK <- sce_protein[,which(sce_protein$celltype == "T_NK")]
```


# T/NK cell subclustering

## Calculate UMAP on all TNK cells

Use only relevant markers. Exclude LAG-3 (did not work well).

```{r TNK-umap}
markers_include_umap <- c("CD38", "CD27", "CD7", "CD45RA", "CD3", "PD1", "Tim-3", "Granzyme_B", "TCF7", "CD45RO", "FOXP3", "ICOS", "CD8a", "GITR", "CD4")

sce_TNK <- runUMAP(sce_TNK, exprs_values = "exprs", subset_row = markers_include_umap)

#sce_TNK <- runUMAP(sce_TNK, dimred = "corrected", name = "UMAP_mnnCorrected", subset_row = markers_include_umap)
```

### Visualize features on UMAP

```{r TNK-umap-features, fig.width=8, fig.height=7}

dittoDimPlot(sce_TNK, var = "sample", reduction.use = "UMAP", size = 0.5, 
                               color.panel = metadata(sce_TNK)$colour_vectors$sample, opacity = 0.5) 

dittoDimPlot(sce_TNK, var = "sample", reduction.use = "UMAP", size = 0.5, 
                               color.panel = metadata(sce_TNK)$colour_vectors$sample, split.by = "sample", opacity = 0.5) 

dittoDimPlot(sce_TNK, var = "TIG", reduction.use = "UMAP", size = 0.5, 
                               color.panel = metadata(sce_TNK)$colour_vectors$TIG)

dittoDimPlot(sce_TNK, var = "TLS", reduction.use = "UMAP", size = 0.5, 
                               color.panel = metadata(sce_TNK)$colour_vectors$TLS)
```

### Visualize marker expression on UMAP

```{r TNK-umap-markerexpression, fig.width=35, fig.height=25, message=FALSE, warning=FALSE}

# Plot expression of all markers
p.list <- list()
for(i in markers_include_umap){
  p.list[[i]] <- plotReducedDim(sce_TNK, colour_by = i, by_exprs_values = "scaled", 
                          point_size = 0.5, point_alpha = 1, shape_by = "panel_type", dimred = "UMAP",)  + scale_colour_gradient2(name = i, low = "#2166ac", mid = "white", high = "#b2182b", limits = c(-3, 3))
}
plot_grid(plotlist = p.list, ncol = 5)
```

## Run graph-based clustering for TNK subset (Phenograph)

Display clusters on UMAP.

```{r TNK-clustering-umap, fig.width=8, fig.height=8}
cluster_counts_TNK <- subset(assay(sce_TNK, "normalized"), rownames(assay(sce_TNK, "normalized")) %in% markers_include_umap)

g <- buildSNNGraph(cluster_counts_TNK, k=30)
clust <- igraph::cluster_louvain(g)$membership
table(clust)

#add cluster information to sce object
sce_TNK$TNK_cluster <- as.factor(clust)

#plot on umap
dittoDimPlot(sce_TNK, var = "TNK_cluster", reduction.use = "UMAP", size = 0.2, do.label = TRUE, opacity = 0.8)
dittoDimPlot(sce_TNK, var = "TNK_cluster", reduction.use = "UMAP", size = 0.2, split.by = "TNK_cluster", opacity = 0.2)
```

### Plot mean expression for each cluster

```{r TNK-clusters-expression-heatmap, fig.width=5, fig.height=6}
clustermean <- aggregateAcrossCells(sce_TNK, ids = DataFrame(cluster = sce_TNK$TNK_cluster), average = TRUE)
colnames(clustermean) <- clustermean$cluster

# Center and scale raw counts
assay(clustermean, "scaled") <- t(scale(t(assay(clustermean, "counts"))))

# Normalize raw counts
norm_counts <- t(apply(assay(clustermean, "counts"), 1, function(x)(x-min(x))/(max(x)-min(x))))
assay(clustermean, "normalized", withDimnames = FALSE) <- norm_counts

## Plot only relevant markers
include_plot <- markers_include_umap

## Plot
# Z-scaled
dittoHeatmap(clustermean[include_plot,], assay = "scaled", scale = "none",
            heatmap.colors = magma(100), 
            cluster_cols = TRUE, show_colnames = TRUE,
            breaks = seq(-3, 3, length.out = 101))

# Normalized (0-1)
dittoHeatmap(clustermean[include_plot,], assay = "normalized", scale = "none",
            heatmap.colors = viridis(100), 
            cluster_cols = TRUE, show_colnames = TRUE)
```

### Plot cluster proportions by sample and sample proportions by cluster

```{r cluster-proportions}
props <- as.data.frame(prop.table(table(sce_TNK$TNK_cluster, sce_TNK$sample), margin = 1))
colnames(props) <- c("cluster", "sample", "proportion")

ggplot(props, aes(cluster, y = proportion, fill=sample)) +
  geom_bar(stat="identity")+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))+
  ggtitle("TNK sample proportions by cluster")

props <- as.data.frame(prop.table(table(sce_TNK$sample, sce_TNK$TNK_cluster), margin = 1))
colnames(props) <- c("cluster", "sample", "proportion")

ggplot(props, aes(cluster, y = proportion, fill=sample)) +
  geom_bar(stat="identity")+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))+
  ggtitle("TNK cluster proportions by sample")
```

### Exclude cluster 9 and plot mean expression for each cluster

```{r TNK-clusters-expression-heatmap-woCluster9, fig.width=5, fig.height=6}
clustermean_wo9 <- clustermean[,-9]

# Center and scale raw counts
assay(clustermean_wo9, "scaled") <- t(scale(t(assay(clustermean_wo9, "counts"))))

# Normalize raw counts
norm_counts <- t(apply(assay(clustermean_wo9, "counts"), 1, function(x)(x-min(x))/(max(x)-min(x))))
assay(clustermean_wo9, "normalized", withDimnames = FALSE) <- norm_counts

## Plot
# Z-scaled
dittoHeatmap(clustermean_wo9[include_plot,], assay = "scaled", scale = "none",
            heatmap.colors = magma(100), 
            cluster_cols = TRUE, show_colnames = TRUE,
            breaks = seq(-3, 3, length.out = 101))

# Normalized (0-1)
dittoHeatmap(clustermean_wo9[include_plot,], assay = "normalized", scale = "none",
            heatmap.colors = viridis(100), 
            cluster_cols = TRUE, show_colnames = TRUE,)
```

### Assign cell type labels

Manual cell type assignment based on canonical marker expression.

```{r TNK-labels, fig.width=8, fig.height=7}
labels <- read.csv(paste0(getwd(), "/data/ProteinPanel_TNK_clusters.csv"))

label_list <- labels$class[match(sce_TNK$TNK_cluster, labels$cluster)]
sce_TNK$subtype <- label_list

# Plot on UMAP
dittoDimPlot(sce_TNK, var = "subtype", reduction.use = "UMAP", size = 0.5, opacity = 0.5, do.label = TRUE)
```

```{r TNK-proportions, fig.width=6, fig.height=4}
#Plot subtype proportions per sample
tab <- as.data.frame(prop.table(table(sce_TNK$sample, sce_TNK$subtype), margin = 1))
colnames(tab) <- c("sample", "subtype", "proportion")

ggplot(tab, aes(sample, y = proportion, fill=subtype)) +
  geom_bar(stat="identity")+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))
```


### Remove "T_AllHigh" cluster for further analysis

This cluster is present in only one sample and appears to be an artefact rather than a true cluster.

```{r}
sce_TNK <- sce_TNK[,which(sce_TNK$subtype != "T_AllHigh")]
```

## Plot expression heatmap with celltype labels

```{r expression_heatmap_celltype_labels}
clustermean <- aggregateAcrossCells(sce_TNK, ids = DataFrame(cluster = sce_TNK$TNK_cluster), average = TRUE)
colnames(clustermean) <- clustermean$cluster

# Normalize raw counts
norm_counts <- t(apply(assay(clustermean, "counts"), 1, function(x)(x-min(x))/(max(x)-min(x))))
assay(clustermean, "normalized", withDimnames = FALSE) <- norm_counts

dittoHeatmap(clustermean[include_plot,], assay = "normalized", scale = "none",
            heatmap.colors = viridis(100),
            cluster_cols = TRUE, show_colnames = TRUE, annot.by = "subtype")

```


# Compare subtype proportions per TIG

Including TLS and pre-TLS.
```{r subtype-proportions-TIG-withTLS, fig.width=10, fig.height=4}
tab <- as.data.frame(prop.table(table(sce_TNK$sample, sce_TNK$subtype), margin = 1))
colnames(tab) <- c("sample", "subtype", "proportion")
TIG2 <- c("TBB111", "TBB129", "TBB165", "TBB171", "TBB184", "TBB338")
tab$TIG <- ifelse(tab$sample %in% TIG2, "TIG2", "TIG3")

ggplot(tab, aes(x= TIG, y = proportion, color=TIG))+
  geom_boxplot()+
  geom_point()+
  facet_wrap(~subtype, scales="free", ncol=6, strip.position="top")+
  theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Of total [%]")+
  stat_compare_means(method = "wilcox.test", aes(label=..p.format..))
```

Excluding TLS and pre-TLS.
```{r subtype-proportions-TIG-woTLS, fig.width=10, fig.height=4}
sce_TNK_woTLS <- sce_TNK[,which(sce_TNK$TLS == "no")]

tab <- as.data.frame(prop.table(table(sce_TNK_woTLS$sample, sce_TNK_woTLS$subtype), margin = 1))
colnames(tab) <- c("sample", "subtype", "proportion")
TIG2 <- c("TBB111", "TBB129", "TBB165", "TBB171", "TBB184", "TBB338")
tab$TIG <- ifelse(tab$sample %in% TIG2, "TIG2", "TIG3")

ggplot(tab, aes(x= TIG, y = proportion, color=TIG))+
  geom_boxplot()+
  geom_point()+
  facet_wrap(~subtype, scales="free", ncol=6, strip.position="top")+
  theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Of total [%]")+
  stat_compare_means(method = "wilcox.test", aes(label=..p.format..))
```

## Percentage of CD8 that belongs to the PD1+ cluster

Excluding TLS and pre-TLS.

```{r CD8PD1-proportions-TIG-woTLS, fig.width=4, fig.height=4}
tab <- as.matrix(prop.table(table(sce_TNK_woTLS$sample, sce_TNK_woTLS$subtype), margin = 1))
CD8_PD1_perc <- as.data.frame(tab[,"CD8_PD1"]/(tab[,"CD8"]+tab[,"CD8_PD1"]))
colnames(CD8_PD1_perc) <- "CD8_PD1_perc"
CD8_PD1_perc$sample <- rownames(CD8_PD1_perc)
CD8_PD1_perc$TIG <- ifelse(CD8_PD1_perc$sample %in% TIG2, "TIG2", "TIG3")

ggplot(CD8_PD1_perc, aes(x= TIG, y = CD8_PD1_perc, color=TIG))+
  geom_boxplot()+
  geom_point()+
  theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("CD8_PD1 of total CD8+ [%]")+
  stat_compare_means(method = "wilcox.test", aes(label=..p.format..))
```

Comparison of TLS, pre-TLS and non-TLS.

```{r CD8PD1-proportions-TLS-comparison, fig.width=4, fig.height=4, message = FALSE}
tab <- as.matrix(prop.table(table(sce_TNK$ImageNumber, sce_TNK$subtype), margin = 1))
CD8_PD1_perc <- as.data.frame(tab[,"CD8_PD1"]/(tab[,"CD8"]+tab[,"CD8_PD1"]))
colnames(CD8_PD1_perc) <- "CD8_PD1_perc"
CD8_PD1_perc$ImageNumber <- rownames(CD8_PD1_perc)

# Retrieve TLS status for each image
img_TLS <- as.data.frame(table(sce_TNK$ImageNumber, sce_TNK$TLS))
colnames(img_TLS) <- c("ImageNumber", "TLS", "cellnumber_image")
img_TLS <- img_TLS %>% filter(cellnumber_image > 0)
img_TLS$cellnumber_image <- NULL

tab.comb <- merge(CD8_PD1_perc, img_TLS,by = "ImageNumber")

TLS_comparisons <- list( c("no", "pre"), c("no", "yes"), c("pre", "yes") )
ggplot(tab.comb, aes(x= TLS, y = CD8_PD1_perc))+
  geom_boxplot()+
  geom_point()+
  theme(axis.text.x=element_text(angle = 90, hjust=1),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("CD8_PD1 of all CD8+ [%]")+
  xlab("TLS classification of the ROI")+
  stat_compare_means(method = "wilcox.test", aes(label=..p.format..), comparison = TLS_comparisons)
```


## Percentage of CD4 that belongs to the PD1+ cluster

Excluding TLS and pre-TLS.

```{r CD4PD1-proportions-TIG-woTLS, fig.width=4, fig.height=4}
tab <- as.matrix(prop.table(table(sce_TNK_woTLS$sample, sce_TNK_woTLS$subtype), margin = 1))
CD4_PD1_perc <- as.data.frame(tab[,"CD4_PD1"]/(tab[,"CD4"]+tab[,"CD4_PD1"]))
colnames(CD4_PD1_perc) <- "CD4_PD1_perc"
CD4_PD1_perc$sample <- rownames(CD4_PD1_perc)
CD4_PD1_perc$TIG <- ifelse(CD4_PD1_perc$sample %in% TIG2, "TIG2", "TIG3")

ggplot(CD4_PD1_perc, aes(x= TIG, y = CD4_PD1_perc, color=TIG))+
  geom_boxplot()+
  geom_point()+
  theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("CD4_PD1 of all CD4+ [%]")+
  stat_compare_means(method = "wilcox.test", aes(label=..p.format..))
```

Comparison of TLS, pre-TLS and non-TLS.

```{r CD4PD1-proportions-TLS-comparison, fig.width=4, fig.height=4, message = FALSE}
tab <- as.matrix(prop.table(table(sce_TNK$ImageNumber, sce_TNK$subtype), margin = 1))
CD4_PD1_perc <- as.data.frame(tab[,"CD4_PD1"]/(tab[,"CD4"]+tab[,"CD4_PD1"]))
colnames(CD4_PD1_perc) <- "CD4_PD1_perc"
CD4_PD1_perc$ImageNumber <- rownames(CD4_PD1_perc)

# Retrieve TLS status for each image
img_TLS <- as.data.frame(table(sce_TNK$ImageNumber, sce_TNK$TLS))
colnames(img_TLS) <- c("ImageNumber", "TLS", "cellnumber_image")
img_TLS <- img_TLS %>% filter(cellnumber_image > 0)
img_TLS$cellnumber_image <- NULL

tab.comb <- merge(CD4_PD1_perc, img_TLS,by = "ImageNumber")

TLS_comparisons <- list( c("no", "pre"), c("no", "yes"), c("pre", "yes") )
ggplot(tab.comb, aes(x= TLS, y = CD4_PD1_perc))+
  geom_boxplot()+
  geom_point()+
  theme(axis.text.x=element_text(angle = 90, hjust=1),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("CD4_PD1 of all CD4+ [%]")+
  xlab("TLS classification of the ROI")+
  stat_compare_means(method = "wilcox.test", aes(label=..p.format..), comparison = TLS_comparisons)
```

## Within-sample variation

Excluding TLS and pre-TLS.

```{r TNK-proportions-sample-variation, fig.width=8, fig.height=6}
#Plot subtype proportions per sample
tab <- as.data.frame(prop.table(table(sce_TNK$ImageNumber, sce_TNK$subtype), margin = 1))
colnames(tab) <- c("ImageNumber", "subtype", "proportion")

ROI_info <- read.csv("data/ROI_info_HIER.csv")
colnames(ROI_info)[8] <- "ImageNumber"
tab.comb <- merge(tab, ROI_info, by = "ImageNumber")
tab.comb<- tab.comb %>% filter(TLS == "no")

## reorder patient levels
tab.comb$sample <- as.factor(tab.comb$sample)
old.levels <- levels(tab.comb$sample)
tab.comb$sample <- factor(tab.comb$sample, levels =  old.levels[c(4:8, 12, 1:3, 9:11)])

ggplot(tab.comb, aes(ImageNumber, y = proportion, fill=subtype)) +
  geom_bar(stat="identity")+
  facet_wrap(~sample, scales="free_x", ncol = 6)+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))
```

```{r CD4PD1-proportions-sample-variation, fig.width=6, fig.height=4}
tab <- as.matrix(prop.table(table(sce_TNK_woTLS$ImageNumber, sce_TNK_woTLS$subtype), margin = 1))
CD4_PD1_perc <- as.data.frame(tab[,"CD4_PD1"]/(tab[,"CD4"]+tab[,"CD4_PD1"]))
colnames(CD4_PD1_perc) <- "CD4_PD1_perc"
CD4_PD1_perc$ImageNumber <- rownames(CD4_PD1_perc)
CD4_PD1_perc <- CD4_PD1_perc[!is.na(CD4_PD1_perc$CD4_PD1_perc),]

tab.comb <- merge(CD4_PD1_perc, tab.comb, by = "ImageNumber")

ggplot(tab.comb, aes(ImageNumber, y = CD4_PD1_perc/6)) +
  geom_bar(stat="identity", aes(fill = TIG))+
  facet_wrap(~sample, scales="free_x", ncol = 6)+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_blank(),
        axis.ticks.x = element_blank(), panel.border = element_rect(colour = "black", fill=NA, size=1))+
  ylab("Proportion CD4_PD1 of all CD4+")+
  ggtitle("CD4_PD1 proportion per ROI")

# Calculate coefficient of variation
cv_CD4PD1 <- tab.comb %>% filter(subtype == "CD4_PD1") %>% group_by(sample) %>% summarize(cv = (sd(CD4_PD1_perc)/mean(CD4_PD1_perc))*100)

# Mean CV including the samples with no CD4_PD1
cv_CD4PD1[is.na(cv_CD4PD1)] <- 0 
meanCV_CD4PD1_incl <- mean(cv_CD4PD1$cv)

# Mean CV excluding the samples with no CD4_PD1
meanCV_CD4PD1_excl <- mean(cv_CD4PD1[cv_CD4PD1$cv != 0,]$cv)
```

```{r CD8PD1-proportions-sample-variation, fig.width=6, fig.height=4}
tab <- as.matrix(prop.table(table(sce_TNK_woTLS$ImageNumber, sce_TNK_woTLS$subtype), margin = 1))
CD8_PD1_perc <- as.data.frame(tab[,"CD8_PD1"]/(tab[,"CD8"]+tab[,"CD8_PD1"]))
colnames(CD8_PD1_perc) <- "CD8_PD1_perc"
CD8_PD1_perc$ImageNumber <- rownames(CD8_PD1_perc)
CD8_PD1_perc <- CD8_PD1_perc[!is.na(CD8_PD1_perc$CD8_PD1_perc),]

tab.comb <- merge(CD8_PD1_perc, tab.comb, by = "ImageNumber")

ggplot(tab.comb, aes(ImageNumber, y = CD8_PD1_perc/6)) +
  geom_bar(stat="identity", aes(fill = TIG))+
  facet_wrap(~sample, scales="free_x", ncol = 6)+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_blank(),
        axis.ticks.x = element_blank(), panel.border = element_rect(colour = "black", fill=NA, size=1))+
  ylab("Proportion CD8_PD1 of all CD8+")+
  ggtitle("CD8_PD1 proportion per ROI")

# Calculate coefficient of variation
cv_CD8PD1 <- tab.comb %>% filter(subtype == "CD8_PD1") %>% group_by(sample) %>% summarize(cv = (sd(CD8_PD1_perc)/mean(CD8_PD1_perc))*100)
meanCV_CD8PD1 <- mean(cv_CD8PD1$cv)
```


# Average marker expression TIG2 vs TIG3

Compare only CD8 T cells.


```{r TIG-marker-comparison-singlecell, fig.width=7, fig.height=4}
#Subset CD8 T cells
sce_CD8 <- sce_TNK[, which(sce_TNK$subtype %in% c("CD8", "CD8_PD1"))]

exprs <- data.frame(`Ki-67` = assay(sce_CD8, "exprs")["Ki-67",],
                    TCF7 = assay(sce_CD8, "exprs")["TCF7",],
                    Granzyme_B = assay(sce_CD8, "exprs")["Granzyme_B",],
                    `Tim-3` = assay(sce_CD8, "exprs")["Tim-3",],
                    GITR = assay(sce_CD8, "exprs")["GITR",],
                    CD38 = assay(sce_CD8, "exprs")["CD38",])
TNK_tab <- cbind(exprs, colData(sce_CD8))
TNK_tab <- pivot_longer(TNK_tab, 1:6, names_to = "marker", values_to = "exprs")

ggplot(TNK_tab, aes(x=TIG, y = exprs, fill = TIG))+
  geom_violin(aes(fill=TIG))+
  geom_boxplot(position=position_dodge(width=0.9), width=0.1, color="black", aes(fill=TIG))+
  facet_wrap(~marker, scale = "free", ncol = 6)+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        panel.background = element_blank(),
        axis.title.x = element_blank())+
  ylab("Marker expression (arcsin-transformed counts)")+
  ggtitle("Single-cell violin plots")
```

Comparison of patient means.

```{r TIG-marker-comparison-patientmeans, fig.width=10, fig.height=5, message=FALSE}

# Patient means
counts <- data.frame(`Ki-67` = assay(sce_CD8, "counts")["Ki-67",],
                    TCF7 = assay(sce_CD8, "counts")["TCF7",],
                    Granzyme_B = assay(sce_CD8, "counts")["Granzyme_B",],
                    `Tim-3` = assay(sce_CD8, "counts")["Tim-3",],
                    GITR = assay(sce_CD8, "counts")["GITR",],
                    CD38 = assay(sce_CD8, "counts")["CD38",])
TNK_tab_2 <- cbind(counts, colData(sce_CD8))
TNK_tab_2 <- pivot_longer(TNK_tab_2, 1:6, names_to = "marker", values_to = "counts")

sample_means <- data.frame(sample = levels(factor(TNK_tab_2$sample)))
for (i in colnames(counts)) {
  TNK.i <- filter(TNK_tab_2, marker == i)
  i.means <- TNK.i %>% group_by(sample) %>% dplyr::summarise(mean = mean(counts))
  sample_means[,i] <- i.means$mean
}

sample_means$TIG <- c(rep("TIG2", 6), rep("TIG3", 6))
sample_means <- pivot_longer(sample_means, 2:7, names_to = "marker", values_to = "counts")

ggplot(sample_means, aes(TIG, counts, color = TIG))+
    geom_boxplot(aes(color=TIG))+
    geom_point(position=position_dodge(width=0.75), aes(group=TIG))+
    facet_wrap(~marker, scales = "free", ncol = 6)+
    stat_compare_means(method = "wilcox.test", aes(label=..p.format..), vjust = 1, label.x.npc = 0.5)+
    theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank(),
        strip.background = element_blank())+
  ggtitle("Patient means")
```


# Proliferation status

Elevated Ki-67 expression in PD1+ T cells (both CD4 and CD8) confirms that these T cell subsets are highly proliferative.


```{r Ki67-expression, fig.width=5, fig.height=4}
sce_TNK.sub <- sce_TNK[,which(sce_TNK$subtype %in% c("CD4", "CD8", "CD4_PD1", "CD8_PD1"))]

dittoPlot(sce_TNK.sub, "Ki-67", group.by = "subtype",
    plots = c("vlnplot"), assay = "exprs")+
  geom_hline(yintercept = 0.7, lty = "dashed")+
  ylab("Arcsinh-counts")+
  theme(legend.position = "none")

TNK_coldata <- data.frame(colData(sce_TNK.sub))
TNK_coldata$Ki67 <- assay(sce_TNK.sub, "scaled")["Ki-67",]

my_comparisons <- list( c("CD4", "CD4_PD1"), c("CD8", "CD8_PD1"))
ggplot(TNK_coldata, aes(subtype, Ki67))+
  geom_violin()+
  stat_compare_means(comparisons = my_comparisons, method = "wilcox.test", aes(label=..p..))

# Define cells with Ki67 arcsinh-counts > 0.5 as proliferating
arcsinh_counts <- assay(sce_TNK, "exprs")
sce_TNK$proliferating <- ifelse(arcsinh_counts["Ki-67",] > 0.7, "yes", "no")
table(sce_TNK$proliferating)

prolif <- prop.table(table(sce_TNK$proliferating, sce_TNK$subtype), margin = 2)
prolif <- as.data.frame(prolif["yes",c("CD4", "CD8", "CD4_PD1", "CD8_PD1")])
colnames(prolif) <- "perc.proliferating"
prolif$CD4CD8 <- c("CD4", "CD8", "CD4", "CD8")
prolif$subtype <- c("PD1-low", "PD1-low", "PD1-high", "PD1-high")

ggplot(prolif, aes(CD4CD8, perc.proliferating, fill = subtype)) +
  geom_col(position=position_dodge())+
  theme(panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1))

  
```

# Add subtypes to full SCE object and save

```{r update-main-SCE}
sce_protein <- readRDS(file = "output/ProteinPanel/sce_ProteinPanel_subtypes.rds")

TNK_ids <- colnames(sce_TNK)
sce_protein[,TNK_ids]$subtype <- sce_TNK$subtype

#Add subtype label also for cluster 9
sce_protein[,which(sce_protein$subtype == "T_NK")]$subtype <- "T_AllHigh"

saveRDS(sce_protein, "output/ProteinPanel/sce_ProteinPanel_subtypes_all.rds")
```

# Myeloid subsetting

Split myeloid cells up into CD11c-high and CD68-high subsets, based on the ratio between the two.

```{r}
# sce_myeloid <- sce_protein[, which(sce_protein$celltype == "myeloid")]
# 
# CD11c_counts <- subset(assay(sce_myeloid, "counts"), rownames(sce_myeloid) == "CD11c")
# CD68_counts <- subset(assay(sce_myeloid, "counts"), rownames(sce_myeloid) == "CD68")
# sce_myeloid$marker_ratio <- as.vector(CD68_counts/CD11c_counts)
# sce_myeloid$subtype <- ifelse(sce_myeloid$marker_ratio >= 1.7, "myeloid_CD68-high", "myeloid_CD11c-high")
# 
# dittoScatterPlot(sce_myeloid, "CD11c", "CD68", opacity = 0.2, color.var = "subtype")
# 
# #Add subtype information to full object
# myeloid_ids <- colnames(sce_myeloid)
# sce_protein[,myeloid_ids]$subtype <- sce_myeloid$subtype
# # 
# # saveRDS(sce_protein, "output/ProteinPanel/sce_ProteinPanel_subtypes_all.rds")

```


# Basic subtype-based analysis

## Marker expression heatmap for all cell subtypes

```{r expression-heatmap-all-subtypes, fig.width=8, fig.height=8}
sce_protein_woTAllHigh <- sce_protein[,which(sce_protein$subtype != "T_AllHigh")]

# Mean marker expression heatmap for all subtypes
subtypemean <- aggregateAcrossCells(sce_protein_woTAllHigh, ids = DataFrame(subtype = sce_protein_woTAllHigh$subtype), average = TRUE)
colnames(subtypemean) <- subtypemean$subtype

# Center and scale raw counts
assay(subtypemean, "scaled") <- t(scale(t(assay(subtypemean, "counts"))))

# Normalize raw counts
norm_counts <- t(apply(assay(subtypemean, "counts"), 1, function(x)(x-min(x))/(max(x)-min(x))))
assay(subtypemean, "normalized", withDimnames = FALSE) <- norm_counts


## Plot
# Z-scaled
dittoHeatmap(subtypemean, assay = "scaled", scale = "none",
            heatmap.colors = magma(100), 
            cluster_cols = TRUE, show_colnames = TRUE,
            breaks = seq(-3, 3, length.out = 101))

# Normalized (0-1)
dittoHeatmap(subtypemean, assay = "normalized", scale = "none",
            heatmap.colors = viridis(100), 
            cluster_cols = TRUE, show_colnames = TRUE)

# aDC markers
# dittoHeatmap(subtypemean[c("LAMP3","CD274", "IDO1", "CD40"),], assay = "normalized", scale = "none",
#             heatmap.colors = viridis(100), 
#             cluster_cols = TRUE, show_colnames = TRUE, cluster_rows = FALSE)
```

### Violinplots of marker expression

```{r violinplots-marker-celltype}
cur_sce_scaled <- t(data.frame(assay(sce_protein_woTAllHigh[c("CD274", "IDO1", "CD40"),], "scaled", withDimnames = TRUE)))
cur_sce_scaled <- data.frame(cur_sce_scaled)
colData <- data.frame(colData(sce_protein_woTAllHigh))
cur_sce_scaled <- cbind(cur_sce_scaled, colData)
cur_sce_scaled <- pivot_longer(cur_sce_scaled, 1:3, names_to = "marker", values_to = "scaled_counts")

ggplot(cur_sce_scaled, aes(x = subtype, y = scaled_counts))+
  geom_violin()+
  facet_wrap(~marker, ncol = 1)


# dittoPlot(sce_protein, "CD274", group.by = "subtype", plots = "vlnplot", assay = "scaled", vlnplot.width = 1.5)
# dittoPlot(sce_protein, "IDO1", group.by = "celltype", plots = "vlnplot", assay = "scaled", vlnplot.width = 1.5)
# dittoPlot(sce_protein, "CD274", group.by = "celltype", plots = "vlnplot", assay = "scaled", vlnplot.width = 1.5)

```


## TLS preference of celltypes

```{r TLS-cellsubtype-preference-timeline}
# Plot as simple timeline
tab <- as.data.frame(prop.table(table(sce_protein$TLS, sce_protein$subtype), margin = 1))
colnames(tab) <- c("TLS", "subtype", "proportion")
tab$TLS <- as.factor(tab$TLS)

ggplot(tab, aes(TLS, proportion, group = 1))+
  geom_line()+
  geom_point()+
  facet_wrap(~subtype, scales = "fixed")
```

```{r TLS-cellsubtype-preference-boxplots, fig.width=12, fig.height=14}
# Boxplots (individual ROIs as points): subtype proportions
tab2 <- as.data.frame(prop.table(table(sce_protein$ImageNumber, sce_protein$subtype), margin = 1))
colnames(tab2) <- c("ImageNumber", "subtype", "proportion")
tab3 <- as.data.frame(table(sce_protein$ImageNumber, sce_protein$TLS))
colnames(tab3) <- c("ImageNumber", "TLS", "cellnumber")
tab3 <- tab3 %>% filter(cellnumber > 0)
tab.comb <- merge(tab2, tab3, by = "ImageNumber")

my_comparisons <- list( c("no", "pre"), c("no", "yes"), c("pre", "yes") )

ggplot(tab.comb, aes(TLS, proportion))+
  geom_boxplot()+
  geom_point()+
  facet_wrap(~subtype, scales = "free")+
    theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Of all stromal [%]")+
  stat_compare_means(comparisons = my_comparisons, method = "wilcox.test", aes(label=..p.signif..))
```

```{r TLS-cellsubtype-number-boxplots, fig.width=12, fig.height=14}

# Boxplots (individual ROIs as points): total celltype number
tab2 <- as.data.frame(table(sce_protein$ImageNumber, sce_protein$subtype), margin = 1)
colnames(tab2) <- c("ImageNumber", "subtype", "cellnumber")
tab3 <- as.data.frame(table(sce_protein$ImageNumber, sce_protein$TLS))
colnames(tab3) <- c("ImageNumber", "TLS", "cellnumber_image")
tab3 <- tab3 %>% filter(cellnumber_image > 0)
tab.comb <- merge(tab2, tab3, by = "ImageNumber")

ggplot(tab.comb, aes(TLS, cellnumber))+
  geom_boxplot()+
  geom_point()+
  stat_compare_means(comparisons = my_comparisons, label = "p.signif")+
  facet_wrap(~subtype, scales = "free", ncol=4)+
    theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())

```

## TIG comparison of subtype frequency
Excluding TLS and pre-TLS.

```{r subtype-frequency-TIG, fig.width=12, fig.height=10}
sce_protein_woTLS <- sce_protein[,which(sce_protein$TLS == "no")]
tab <- as.data.frame(prop.table(table(sce_protein_woTLS$sample, sce_protein_woTLS$subtype), margin = 1))
colnames(tab) <- c("sample", "subtype", "proportion")
TIG2 <- c("TBB111", "TBB129", "TBB165", "TBB171", "TBB184", "TBB338")
tab$TIG <- ifelse(tab$sample %in% TIG2, "TIG2", "TIG3")

ggplot(tab, aes(x= TIG, y = proportion, color=TIG))+
  geom_boxplot()+
  geom_point()+
  facet_wrap(~subtype, scales="free", ncol=7, strip.position="top")+
  theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Of total [%]")+
  stat_compare_means(method = "wilcox.test", aes(label=..p.format..))
```

## Within sample variation of main celltypes

Excluding TLS and pre-TLS.

Use CV (coefficient of variation) and PV (proportional variability).
PV is less susceptible to large outliers, and handles zeros better.

```{r subtype-proportions-sample-variation, fig.width=14, fig.height=25}
#Plot subtype proportions per sample
tab <- as.data.frame(prop.table(table(sce_protein$ImageNumber, sce_protein$celltype), margin = 1))
colnames(tab) <- c("ImageNumber", "celltype", "proportion")

tab.comb <- merge(tab, ROI_info, by = "ImageNumber")
tab.comb<- tab.comb %>% filter(TLS == "no")

## reorder patient levels
tab.comb$sample <- as.factor(tab.comb$sample)
old.levels <- levels(tab.comb$sample)
tab.comb$sample <- factor(tab.comb$sample, levels =  old.levels[c(4:8, 12, 1:3, 9:11)])

p.list <- list()
for (i in unique(sce_protein$celltype)) {
  tab.i <- tab.comb %>% filter(celltype == i)
p.list[[i]]<- ggplot(tab.i, aes(ImageNumber, y = proportion, fill=TIG)) +
  geom_bar(stat="identity")+
  facet_wrap(~sample, scales="free_x", ncol = 12)+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))+
  ggtitle(i)
}
plot_grid(plotlist = p.list, ncol = 1)

# Calculate mean CVs
cvs <- data.frame(mean_cv = numeric())
for (i in unique(sce_protein$celltype)) {
cv_i <- tab.comb %>% filter(celltype == i) %>% group_by(sample) %>% summarize(cv = (sd(proportion)/mean(proportion))*100)
meanCV_i<- mean(cv_i$cv, na.rm = TRUE)
cvs[i,] <- meanCV_i
}
```

```{r subtype-proportions-CV, fig.width=5, fig.height=4}
# Sort by mean CV and plot
cvs$celltype <- as.character(rownames(cvs))
ggplot(cvs[order(cvs$mean_cv),], aes(reorder(celltype, mean_cv), y = mean_cv)) +
  geom_bar(stat="identity")+
  theme(axis.text.x = element_text(angle = 90, hjust = 1), axis.title.x = element_blank(),
        panel.background = element_blank())+
  ggtitle("Mean CV of cell type proportions\n(across ROIs of the same sample)")

# Calculate mean PV (Proportional Variability)
pv <- data.frame(mean_pv = numeric())
for (i in unique(sce_protein$celltype)) {
pv_i <- tab.comb %>% filter(celltype == i) %>% group_by(sample) %>% summarize(pv = PV(proportion))
meanPV_i<- mean(pv_i$pv, na.rm = TRUE)
pv[i,] <- meanPV_i
}

# Sort by mean PV and plot
pv$celltype <- as.character(rownames(pv))
ggplot(pv[order(pv$mean_pv),], aes(reorder(celltype, mean_pv), y = mean_pv)) +
  geom_bar(stat="identity")+
  theme(axis.text.x = element_text(angle = 90, hjust = 1), axis.title.x = element_blank(),
        panel.background = element_blank())+
  ggtitle("Mean PV of cell type proportions\n(across ROIs of the same sample)")

```

### For tumor subtypes only.

```{r tumor-proportions-sample-variation, fig.width=14, fig.height=12}
sce_tumor <- sce_protein[,which(sce_protein$celltype == "tumor")]
tab <- as.data.frame(prop.table(table(sce_tumor$ImageNumber, sce_tumor$subtype), margin = 1))
colnames(tab) <- c("ImageNumber", "subtype", "proportion")

tab.comb <- merge(tab, ROI_info, by = "ImageNumber")
tab.comb<- tab.comb %>% filter(TLS == "no")

## reorder patient levels
tab.comb$sample <- as.factor(tab.comb$sample)
old.levels <- levels(tab.comb$sample)
tab.comb$sample <- factor(tab.comb$sample, levels =  old.levels[c(4:8, 12, 1:3, 9:11)])

p.list <- list()
for (i in unique(sce_tumor$subtype)) {
  tab.i <- tab.comb %>% filter(subtype == i)
p.list[[i]]<- ggplot(tab.i, aes(ImageNumber, y = proportion, fill=TIG)) +
  geom_bar(stat="identity")+
  facet_wrap(~sample, scales="free_x", ncol = 12)+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))+
  ggtitle(i)
}
plot_grid(plotlist = p.list, ncol = 1)

# Calculate mean CVs
cvs <- data.frame(mean_cv = numeric())
for (i in unique(sce_tumor$subtype)) {
cv_i <- tab.comb %>% filter(subtype == i) %>% group_by(sample) %>% summarize(cv = (sd(proportion)/mean(proportion))*100)
meanCV_i<- mean(cv_i$cv)
cvs[i,] <- meanCV_i
}

# Calculate mean PV (Proportional Variability)
pv <- data.frame(mean_pv = numeric())
for (i in unique(sce_tumor$subtype)) {
pv_i <- tab.comb %>% filter(subtype == i) %>% group_by(sample) %>% summarize(pv = PV(proportion))
meanPV_i<- mean(pv_i$pv, na.rm = TRUE)
pv[i,] <- meanPV_i
}
```

```{r tumor-proportions-CV, fig.width=4, fig.height=4}
# Sort by mean CV and plot
cvs$subtype <- as.character(rownames(cvs))
ggplot(cvs[order(cvs$mean_cv),], aes(reorder(subtype, mean_cv), y = mean_cv)) +
  geom_bar(stat="identity")+
  theme(axis.text.x = element_text(angle = 90, hjust = 1), axis.title.x = element_blank(),
        panel.background = element_blank())+
  ggtitle("Mean CV of cell type proportions\n(across ROIs of the same sample)")

# Sort by mean PV and plot
pv$subtype <- as.character(rownames(pv))
ggplot(pv[order(pv$mean_pv),], aes(reorder(subtype, mean_pv), y = mean_pv)) +
  geom_bar(stat="identity")+
  theme(axis.text.x = element_text(angle = 90, hjust = 1), axis.title.x = element_blank(),
        panel.background = element_blank())+
  ggtitle("Mean PV of cell type proportions\n(across ROIs of the same sample)")
```

### For T/NK subtypes only.

```{r TNK-proportions-variation, fig.width=14, fig.height=12}
sce_TNK <- sce_protein[,which(sce_protein$celltype == "T_NK")]
tab <- as.data.frame(prop.table(table(sce_TNK$ImageNumber, sce_TNK$subtype), margin = 1))
colnames(tab) <- c("ImageNumber", "subtype", "proportion")

tab.comb <- merge(tab, ROI_info, by = "ImageNumber")
tab.comb<- tab.comb %>% filter(TLS == "no")

## reorder patient levels
tab.comb$sample <- as.factor(tab.comb$sample)
old.levels <- levels(tab.comb$sample)
tab.comb$sample <- factor(tab.comb$sample, levels =  old.levels[c(4:8, 12, 1:3, 9:11)])

p.list <- list()
for (i in unique(sce_TNK$subtype)) {
  tab.i <- tab.comb %>% filter(subtype == i)
p.list[[i]]<- ggplot(tab.i, aes(ImageNumber, y = proportion, fill=TIG)) +
  geom_bar(stat="identity")+
  facet_wrap(~sample, scales="free_x", ncol = 12)+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))+
  ggtitle(i)
}
plot_grid(plotlist = p.list, ncol = 1)

# Calculate mean CVs
cvs <- data.frame(mean_cv = numeric())
for (i in unique(sce_TNK$subtype)) {
cv_i <- tab.comb %>% filter(subtype == i) %>% group_by(sample) %>% summarize(cv = (sd(proportion)/mean(proportion))*100)
meanCV_i<- mean(cv_i$cv, na.rm = TRUE)
cvs[i,] <- meanCV_i
}

# Calculate mean PV (Proportional Variability)
pv <- data.frame(mean_pv = numeric())
for (i in unique(sce_TNK$subtype)) {
pv_i <- tab.comb %>% filter(subtype == i) %>% group_by(sample) %>% summarize(pv = PV(proportion))
meanPV_i<- mean(pv_i$pv, na.rm = TRUE)
pv[i,] <- meanPV_i
}
```

```{r TNK-proportions-CV, fig.width=5, fig.height=4}
# Sort by mean CV and plot
cvs$subtype <- as.character(rownames(cvs))
ggplot(cvs[order(cvs$mean_cv),], aes(reorder(subtype, mean_cv), y = mean_cv)) +
  geom_bar(stat="identity")+
  theme(axis.text.x = element_text(angle = 90, hjust = 1), axis.title.x = element_blank(),
        panel.background = element_blank())+
  ggtitle("Mean CV of cell type proportions\n(across ROIs of the same sample)")

# Sort by mean PV and plot
pv$subtype <- as.character(rownames(pv))
ggplot(pv[order(pv$mean_pv),], aes(reorder(subtype, mean_pv), y = mean_pv)) +
  geom_bar(stat="identity")+
  theme(axis.text.x = element_text(angle = 90, hjust = 1), axis.title.x = element_blank(),
        panel.background = element_blank())+
  ggtitle("Mean PV of cell type proportions\n(across ROIs of the same sample)")
```
