==> analysis/10_RNApanel_TNK_cytokines.Rmd <==
---
title: "10_RNApanel_TNK_cytokines"
author: "SandraTietscher"
date: "2020-10-30"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction

```{r setup, include=FALSE, message=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
```

This script defines CXCL13-expressing T cells and CSF1-expressing T/NK cells and looks into them in more detail.


# Load libraries and data

```{r load-libraries-and-data, message=FALSE}
library(SingleCellExperiment)
library(dplyr)
library(ggplot2)
library(scater)
library(tidyr)
library(ggpubr)
library(dittoSeq)
library(Seurat)
library(RANN)
library(stringr)

sce_RNA <- readRDS(file = "output/RNApanel/sce_RNApanel_celltypes_cytokines.rds")
TIG2 <- c("TBB111", "TBB129", "TBB165", "TBB171", "TBB184", "TBB338")
```

# Assign new subtypes to CXCL13-expressing CD4 and CD8 T cells

```{r}
sce_RNA$subtype <- sce_RNA$celltype
sce_RNA$subtype <- ifelse(sce_RNA$celltype == "T_CD4" & sce_RNA$CXCL13_mRNA_expression == 1, "T_CD4_CXCL13", sce_RNA$subtype)
sce_RNA$subtype <- ifelse(sce_RNA$celltype == "T_CD8" & sce_RNA$CXCL13_mRNA_expression == 1, "T_CD8_CXCL13", sce_RNA$subtype)
```

## CXCL13+ T cell proportions

### Across TLS states

```{r CXCL13-frequency-TLS}
sce_CD8 <- sce_RNA[,which(sce_RNA$celltype == "T_CD8")]
sce_CD4 <- sce_RNA[,which(sce_RNA$celltype == "T_CD4")]


# Boxplots (individual ROIs as points)
tab_CD8 <- as.data.frame(prop.table(table(sce_CD8$ImageNumber, sce_CD8$subtype), margin = 1)) %>% filter(Var2 == "T_CD8_CXCL13")
frequency <- as.data.frame(table(sce_CD8$ImageNumber, sce_CD8$CXCL13_mRNA_expression)) %>% filter(Var2 == 1)
tab_CD8$frequency <- frequency[,3]

tab_CD4 <- as.data.frame(prop.table(table(sce_CD4$ImageNumber, sce_CD4$subtype), margin = 1)) %>% filter(Var2 == "T_CD4_CXCL13")
frequency <- as.data.frame(table(sce_CD4$ImageNumber, sce_CD4$CXCL13_mRNA_expression)) %>% filter(Var2 == 1)
tab_CD4$frequency <- frequency[,3]

tab_comb <- rbind(tab_CD8, tab_CD4)
colnames(tab_comb) <- c("ImageNumber", "subtype", "proportion", "frequency")


tab <- as.data.frame(table(sce_CD8$ImageNumber, sce_CD8$TLS))
colnames(tab) <- c("ImageNumber", "TLS", "cellnumber_image")
tab <- tab %>% filter(cellnumber_image > 0)
tab$cellnumber_image <- NULL
tab_comb2 <- merge(tab_comb, tab, by = "ImageNumber")

my_comparisons <- list( c("no", "pre"), c("no", "yes"), c("pre", "yes") )

ggplot(tab_comb2, aes(TLS, frequency))+
  geom_boxplot()+
  geom_point()+
  stat_compare_means(comparisons = my_comparisons, label = "p.format", method = "wilcox")+
  facet_wrap(~subtype, scales = "fixed")+
    theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Frequency per ROI")
```

```{r CXCL13-proportion-TLS}
ggplot(tab_comb2, aes(TLS, proportion))+
  geom_boxplot()+
  geom_point()+
  stat_compare_means(comparisons = my_comparisons, label = "p.format", method = "wilcox")+
  facet_wrap(~subtype, scales = "fixed")+
    theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Proportion of all CD8+/CD4+ T cells per ROI")
```

### Across samples/TIGs (for non-TLS only)

```{r CXCL13-proportion-TIG}
sce_CD8_TLSex <- sce_CD8[,which(sce_CD8$TLS == "no")]
sce_CD4_TLSex <- sce_CD4[,which(sce_CD4$TLS == "no")]


# Boxplots (individual ROIs as points)
tab_CD8 <- as.data.frame(prop.table(table(sce_CD8_TLSex$sample, sce_CD8_TLSex$subtype), margin = 1)) %>% filter(Var2 == "T_CD8_CXCL13")
frequency <- as.data.frame(table(sce_CD8_TLSex$sample, sce_CD8_TLSex$subtype)) %>% filter(Var2 == "T_CD8_CXCL13")
tab_CD8$frequency <- frequency[,3]

tab_CD4 <- as.data.frame(prop.table(table(sce_CD4_TLSex$sample, sce_CD4_TLSex$subtype), margin = 1)) %>% filter(Var2 == "T_CD4_CXCL13")
frequency <- as.data.frame(table(sce_CD4_TLSex$sample, sce_CD4_TLSex$subtype)) %>% filter(Var2 == "T_CD4_CXCL13")
tab_CD4$frequency <- frequency[,3]

tab_comb <- rbind(tab_CD8, tab_CD4)
colnames(tab_comb) <- c("sample", "subtype", "proportion", "frequency")
tab_comb$TIG <- ifelse(tab_comb$sample %in% TIG2, "TIG2", "TIG3")


ggplot(tab_comb, aes(TIG, proportion, color = TIG))+
  geom_boxplot()+
  geom_point()+
  stat_compare_means(label = "p.format", method = "wilcox", label.x.npc = "middle")+
  facet_wrap(~subtype, scales = "fixed")+
  theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Proportion out of all CD8+/CD4+ T cells")
```

## Check Ki-67 status

Ki-67 expression cutoff was chosen at 0.3 mean counts per cell.

```{r Ki67-status-CXCL13}
# For CD8+ CXCL13+ T cells
Ki67_exprs <- as.data.frame(assay(sce_CD8, "counts")["Ki-67",])
colnames(Ki67_exprs) <- "Ki67_exprs"
Ki67_exprs$Ki67_status <- ifelse(Ki67_exprs$Ki67_exprs > 0.3, "positive", "negative")
tab <- cbind(Ki67_exprs, colData(sce_CD8))
tab$CXCL13_status <- ifelse(tab$CXCL13_mRNA_expression == 1, "positive", "negative")

# For CD4+ CXCL13+ T cells
Ki67_exprs_CD4 <- as.data.frame(assay(sce_CD4, "counts")["Ki-67",])
colnames(Ki67_exprs_CD4) <- "Ki67_exprs"
Ki67_exprs_CD4$Ki67_status <- ifelse(Ki67_exprs_CD4$Ki67_exprs > 0.3, "positive", "negative")
tab_CD4 <- cbind(Ki67_exprs_CD4, colData(sce_CD4))
tab_CD4$CXCL13_status <- ifelse(tab_CD4$CXCL13_mRNA_expression == 1, "positive", "negative")

tab_comb <- rbind(tab, tab_CD4)

ggplot(tab_comb, aes(x = CXCL13_status, fill = Ki67_status))+
  geom_bar(position="fill")+
  scale_fill_manual(values = c("black", "deepskyblue1"))+
  facet_wrap(~celltype)+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        panel.background = element_blank())+
  ylab("Proportion [%]")
```

# Assign CSF1+ NKT cell subtype

All CSF1-positive T and NK cells (except CXCL13+) will be classified as NKT cells. This is only an approximation, but as close as we can get to defining this cell type.

```{r CSF1-assign-subtype}
sce_RNA$subtype <- ifelse(sce_RNA$subtype %in% c("NK", "T_CD8", "T_CD4") & sce_RNA$CSF1_expression == 1, "CSF1_NKT", sce_RNA$subtype)
```

## Check CSF1_NKT proportions/frequency 

### Across TLS states; proportion out of all cells

```{r CSF1-frequency-TLS, fig.width=3, fig.height=4}
tab <- as.data.frame(prop.table(table(sce_RNA$ImageNumber, sce_RNA$subtype), margin = 1)) %>% filter(Var2 == "CSF1_NKT")
frequency <- as.data.frame(table(sce_RNA$ImageNumber, sce_RNA$subtype)) %>% filter(Var2 == "CSF1_NKT")
tab$frequency <- frequency[,3]
colnames(tab) <- c("ImageNumber", "subtype", "proportion", "frequency")


tab2 <- as.data.frame(table(sce_RNA$ImageNumber, sce_RNA$TLS))
colnames(tab2) <- c("ImageNumber", "TLS", "cellnumber_image")
tab2 <- tab2 %>% filter(cellnumber_image > 0)
tab2$cellnumber_image <- NULL
tab_comb <- merge(tab2, tab, by = "ImageNumber")

my_comparisons <- list( c("no", "pre"), c("no", "yes"), c("pre", "yes") )

ggplot(tab_comb, aes(TLS, frequency))+
  geom_boxplot()+
  geom_point()+
  stat_compare_means(comparisons = my_comparisons, label = "p.format", method = "wilcox")+
  facet_wrap(~subtype, scales = "fixed")+
    theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Frequency per ROI")
```

```{r CSF1-proportion-TLS, fig.width=3, fig.height=4}
ggplot(tab_comb, aes(TLS, proportion))+
  geom_boxplot()+
  geom_point()+
  stat_compare_means(comparisons = my_comparisons, label = "p.format", method = "wilcox")+
  facet_wrap(~subtype, scales = "fixed")+
    theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Proportion of all cells per ROI")
```

### Across samples/TIGs, proportion out of T/NK cells

For non-TLS images only.

```{r cSF1-proportion-TIG, fig.width=3, fig.height=4}
sce_TLSex <- sce_RNA[,which(sce_RNA$TLS == "no")]
sce_TLSex_TNK <- sce_TLSex[,which(sce_TLSex$celltype %in% c("T_CD4", "T_CD8", "NK"))]

# Boxplots (individual ROIs as points)
tab <- as.data.frame(prop.table(table(sce_TLSex_TNK$sample, sce_TLSex_TNK$subtype), margin = 1)) %>% filter(Var2 == "CSF1_NKT")
frequency <- as.data.frame(table(sce_TLSex_TNK$sample, sce_TLSex_TNK$subtype)) %>% filter(Var2 == "CSF1_NKT")
tab$frequency <- frequency[,3]

colnames(tab) <- c("sample", "subtype", "proportion", "frequency")
tab$TIG <- ifelse(tab$sample %in% TIG2, "TIG2", "TIG3")

ggplot(tab, aes(TIG, proportion, color = TIG))+
  geom_boxplot()+
  geom_point()+
  stat_compare_means(label = "p.format", method = "wilcox", label.x.npc = "middle")+
  facet_wrap(~subtype, scales = "fixed")+
    theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Proportion out of T/NK cells")
```

# Compare to 10x proportions
Exclude TLS and pre-TLS

## Compare CSF1_NKT proportions (out of all T/NK cells)

```{r CSF1-proportions-method-comparison}
# Read in Seurat object and calculate NKT proportions per sample
Tcells_10x <- readRDS("data/RNAseq_Tcells.rds")
prop <- as.data.frame(prop.table(table(Tcells_10x$orig.ident, Tcells_10x$metacluster), margin = 1))
colnames(prop) <- c("sample", "metacluster", "RNAseq_proportion")
prop <- prop[prop$metacluster == "NKT",]

# Remove samples that were not measured by IMC
not.measured <- c("TBB011", "TBB330")
prop <- prop[!(prop$sample %in% not.measured),]

# Retrieve IMC proportions (of all T/NK cells)
sce_TNK <- sce_RNA[, which(sce_RNA$celltype %in% c("NK", "T_CD4", "T_CD8"))]
sce_TNK_TLSex <- sce_TNK[,which(sce_TNK$TLS == "no")]
IMC_prop <- as.data.frame(prop.table(table(sce_TNK_TLSex$sample, sce_TNK_TLSex$subtype), margin = 1)) %>% filter(Var2 == "CSF1_NKT")
colnames(IMC_prop) <- c("sample", "subtype", "IMC_proportion")

# Merge
prop.comb <- merge(IMC_prop, prop, by = "sample")
prop.comb$TIG <- ifelse(prop.comb$sample %in% TIG2, "TIG2", "TIG3")

# Test correlation
ggplot(prop.comb, aes(IMC_proportion, RNAseq_proportion)) +
  geom_point(aes(color=TIG))+
  geom_smooth(method = lm, color = "black", se=FALSE)+
  stat_cor(method="spearman")+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  ggtitle("CSF1_NKT proportion out of all T/NK cells - method comparison")

```

## Compare CD4_CXCL13 proportions (out of all T/NK cells)

```{r, CD4_CXCL13-proportions-method-comparison}
# Read in Seurat object and calculate NKT proportions per sample
prop <- as.data.frame(prop.table(table(Tcells_10x$orig.ident, Tcells_10x$metacluster), margin = 1))
colnames(prop) <- c("sample", "metacluster", "RNAseq_proportion")
prop <- prop[prop$metacluster == "CD4_exhausted",]

# Remove samples that were not measured by IMC
not.measured <- c("TBB011", "TBB330")
prop <- prop[!(prop$sample %in% not.measured),]

# Retrieve IMC proportions (of all T/NK cells)
IMC_prop <- as.data.frame(prop.table(table(sce_TNK_TLSex$sample, sce_TNK_TLSex$subtype), margin = 1)) %>% filter(Var2 == "T_CD4_CXCL13")
colnames(IMC_prop) <- c("sample", "subtype", "IMC_proportion")

# Merge
prop.comb <- merge(IMC_prop, prop, by = "sample")
prop.comb$TIG <- ifelse(prop.comb$sample %in% TIG2, "TIG2", "TIG3")

# Test correlation
ggplot(prop.comb, aes(IMC_proportion, RNAseq_proportion)) +
  geom_point(aes(color=TIG))+
  geom_smooth(method = lm, color = "black", se=FALSE)+
  stat_cor(method="spearman")+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  ggtitle("CD4_CXCL13+ proportion out of all T/NK cells - method comparison")

```

## Compare CD8_CXCL13 proportions (out of all T/NK cells)

```{r CD8_CXCL13-proportions-method-comparison}
# Read in Seurat object and calculate NKT proportions per sample
prop <- as.data.frame(prop.table(table(Tcells_10x$orig.ident, Tcells_10x$metacluster), margin = 1))
colnames(prop) <- c("sample", "metacluster", "RNAseq_proportion")
prop <- prop[prop$metacluster == "CD8_exhausted",]

# Remove samples that were not measured by IMC
not.measured <- c("TBB011", "TBB330")
prop <- prop[!(prop$sample %in% not.measured),]

# Retrieve IMC proportions (of all T/NK cells)
IMC_prop <- as.data.frame(prop.table(table(sce_TNK_TLSex$sample, sce_TNK_TLSex$subtype), margin = 1)) %>% filter(Var2 == "T_CD8_CXCL13")
colnames(IMC_prop) <- c("sample", "subtype", "IMC_proportion")

# Merge
prop.comb <- merge(IMC_prop, prop, by = "sample")
prop.comb$TIG <- ifelse(prop.comb$sample %in% TIG2, "TIG2", "TIG3")

# Test correlation
ggplot(prop.comb, aes(IMC_proportion, RNAseq_proportion)) +
  geom_point(aes(color=TIG))+
  geom_smooth(method = lm, color = "black", se=FALSE)+
  stat_cor(method="spearman")+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  ggtitle("CD8_CXCL13+ proportion out of all T/NK cells - method comparison")

```

# CSF1_NKT cells: correlation with myeloid activation/differentiation status 

To see whether myeloid activation occurs close to NKT cells.


## With distance to the next NKT cells as spatial measure

Set 800 as maximum distance.

```{r calculate-distance-CSF1}
max_dist <- 800
cur_sce <- data.frame(colData(sce_RNA))
cellIDs_NKT <- cur_sce[cur_sce[,"subtype"] == "CSF1_NKT",]
new_sce <- data.frame(matrix(ncol=length(colnames(cur_sce))+2, nrow = 0))

# For each image separately: nearest neighbour search

for (j in unique(cur_sce$ImageNumber)) {
  cur_sce_j <- filter(cur_sce, ImageNumber == j)
  NKT_j <- filter(cellIDs_NKT, ImageNumber == j)[,c("Center_X", "Center_Y")]
  
  #For images in which there is no NKT cell --> set distance parameter to max
  if (nrow(NKT_j) == 0) {
    cur_sce_j$NKT_distance <- max_dist
  } else {
    j_coord <- cur_sce_j[,c("Center_X", "Center_Y")]
    nn <- RANN::nn2(NKT_j, j_coord)
    cur_sce_j$NKT_distance <- nn$nn.dists[,1]
    
    #Set distance parameter to max for all cells with distance > max
    cur_sce_j$NKT_distance<- ifelse(cur_sce_j$NKT_distance > max_dist, max_dist, cur_sce_j$NKT_distance)
    cur_sce_j$ImageNumber <- j
  }
  new_sce <- rbind(cur_sce_j, new_sce)
}

plot(density(new_sce$NKT_distance))

# Add NKT_distances to colData of original object
sce_RNA$NKT_distance <- new_sce[order(match(new_sce$cellID, sce_RNA$cellID)),]$NKT_distance
plot(density(sce_RNA[,which(sce_RNA$subtype == "CSF1_NKT")]$NKT_distance))

# Bin the distances
breaks <- seq(from = 0, to = 800, by = 5)
sce_RNA$NKT_distance_bin <- cut(sce_RNA$NKT_distance, breaks = breaks, include.lowest = TRUE, right = TRUE, ordered_result = TRUE)

breaks50 <- seq(from = 0, to = 800, by = 50)
sce_RNA$NKT_distance_bin50 <- cut(sce_RNA$NKT_distance, breaks = breaks50, include.lowest = TRUE, right = TRUE, ordered_result = TRUE)

```

### Use the cytokine-expression status as a proxy for cell activation

```{r myeloid-activation-cytokines-CSF1-distance, fig.width=5, fig.height=5}
sce_myeloid <- sce_RNA[,which(sce_RNA$celltype == "myeloid")]

cytokine_myeloid <- as.data.frame(prop.table(table(sce_myeloid$NKT_distance_bin, sce_myeloid$cytokine), margin = 1)[,2])
cytokine_myeloid$distance <- seq(from = 2.5, to = 800, by = 5)
colnames(cytokine_myeloid)[1] <- "prop_cytokine_expressing"

ggplot(cytokine_myeloid, aes(distance, prop_cytokine_expressing)) +
  geom_point()+
  #geom_smooth(method = "gam", color = "black", se=FALSE)+
  #stat_cor(method="spearman")+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  ylab("Proportion of myeloid cells that express at least one cytokine")+
  xlab("Distance to the next NKT cells (5um bins)")


## Separately for the samples
cytokine_myeloid_sample <- data.frame(matrix(nrow=0, ncol=3))
  
for (i in unique(sce_myeloid$sample)) {
  sce_myeloid_i <- sce_myeloid[,which(sce_myeloid$sample == i)]
  cytokine_myeloid_i <- as.data.frame(prop.table(table(sce_myeloid_i$NKT_distance_bin50, sce_myeloid_i$cytokine), margin = 1)[,2])
  cytokine_myeloid_i$distance <- seq(from = 25, to = 800, by = 50)
  cytokine_myeloid_i$sample <- i
  cytokine_myeloid_sample <- rbind(cytokine_myeloid_sample, cytokine_myeloid_i)
}

colnames(cytokine_myeloid_sample)[1] <- "prop_cytokine_expressing"

ggplot(cytokine_myeloid_sample, aes(distance, prop_cytokine_expressing, color = sample)) +
  geom_line(aes(group=sample))+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  ylab("Proportion of myeloid cells that express at least one cytokine")+
  xlab("Distance to the next NKT cells (50um bins)")  


# Smoothing instead of binning
colData_myeloid <- data.frame(colData(sce_myeloid))

ggplot(colData_myeloid, aes(NKT_distance, as.numeric(cytokine))) +
  geom_smooth(aes(group=sample, color=sample), method = loess)+
  #geom_smooth(method = loess, se = FALSE)+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  coord_cartesian(xlim = c(0,400))+
  ylab("Proportion of myeloid cells that express at least one cytokine")+
  xlab("Distance to the next NKT cell")  
```

#### Control: Same analysis for other cytokines

```{r control-other-cytokines}
# Calculate macrophage activation status based on CD68/CD163/CD204/CCL18 (for all cells, will be subset later)
  # Re-normalize counts
  norm_counts <- t(apply(assay(sce_RNA, "counts"), 1, function(x)(x-min(x))/(quantile(x, 0.99)-min(x))))
  norm_counts <- t(apply(norm_counts, 1, function(x) pmin(x, 1)))
  assay(sce_RNA, "normalized", withDimnames = FALSE) <- norm_counts

  # For each cells, average across the normalized counts for CD68, CD163, CD204 and CCL18
  count_matrix <- assay(sce_RNA, "normalized")[c("CD68", "CD163", "CD204", "CCL18"),]
  sce_RNA$activation_score <- colMeans(count_matrix)

# Initialize binned table
myeloid_distances <- data.frame(cellsOI_distance_bin = cut(seq(from = 5, to = 800,  5), breaks = breaks, include.lowest = TRUE, right = TRUE, ordered_result = TRUE))



# Loop over all cytokines
cytokine_names <- paste0(c("CCL4", "CCL18", "CSF1", "CXCL10", "CXCL13_mRNA", "CCL2", "CCL22", "CXCL9", "CCL17", "CCL5"), "_expression")

for (i in cytokine_names) {
  cur_sce <- data.frame(colData(sce_RNA))
  cellsOI <- cur_sce[cur_sce[,i] == TRUE & cur_sce$celltype %in% c("NK", "T_CD8", "T_CD4"),]
  new_sce <- data.frame(matrix(ncol=length(colnames(cur_sce))+2, nrow = 0))

  # Loop over all images to perform nearest neighbour search
  for (j in unique(cur_sce$ImageNumber)) {
    cur_sce_j <- filter(cur_sce, ImageNumber == j)
    cellsOI_j <- filter(cellsOI, ImageNumber == j)[,c("Center_X", "Center_Y")]
    
    #For images in which there is no cell of interest --> set distance parameter to max
    if (nrow(cellsOI_j) == 0) {
      cur_sce_j$cellsOI_distance <- max_dist
    } else {
      j_coord <- cur_sce_j[,c("Center_X", "Center_Y")]
      nn <- RANN::nn2(cellsOI_j, j_coord)
      cur_sce_j$cellsOI_distance <- nn$nn.dists[,1]
      
      #Set distance parameter to max for all cells with distance > max
      cur_sce_j$cellsOI_distance<- ifelse(cur_sce_j$cellsOI_distance > max_dist, max_dist, cur_sce_j$cellsOI_distance)
      cur_sce_j$ImageNumber <- j
    }
    new_sce <- rbind(cur_sce_j, new_sce)
  }
  
  # Add cellsOI_distances to colData of original object
  sce_RNA$cellsOI_distance <- new_sce[order(match(new_sce$cellID, sce_RNA$cellID)),]$cellsOI_distance

  # Bin the distances
  breaks <- seq(from = 0, to = 800, by = 5)
  sce_RNA$cellsOI_distance_bin <- cut(sce_RNA$cellsOI_distance, breaks = breaks, include.lowest = TRUE, right = TRUE, ordered_result = TRUE)
  
  # Extract the myeloid activation status (based on cytokine expression) in relation to the cellsOI distance
  sce_myeloid <- sce_RNA[,which(sce_RNA$celltype == "myeloid")]
  cytokine_myeloid <- as.data.frame(prop.table(table(sce_myeloid$cellsOI_distance_bin, sce_myeloid$cytokine), margin = 1)[,2])
  colnames(cytokine_myeloid)[1] <- paste0(i, "_activationCytokine")
  cytokine_myeloid$cellsOI_distance_bin <- rownames(cytokine_myeloid)
  
  # Extract the myeloid activation status based on CD68/CD163/CD204/CCL18 expression
  coldata_myeloid <- data.frame(colData(sce_myeloid))
  myeloid_summarized <- coldata_myeloid %>% group_by(cellsOI_distance_bin) %>% summarize(mean_activation_score = mean(activation_score))
  comb <- merge(cytokine_myeloid, myeloid_summarized, by = "cellsOI_distance_bin")
  colnames(comb)[3] <- paste0(i, "_activationScore")
  
  myeloid_distances <- merge(myeloid_distances, comb, by = "cellsOI_distance_bin")
}



## Myeloid cell activation measured as cytokine expression status
myeloid_distances$cellsOI_distance_bin <- sub("\\(", "", myeloid_distances$cellsOI_distance_bin)
myeloid_distances$cellsOI_distance_bin <- as.numeric(sub(",.*", "", myeloid_distances$cellsOI_distance_bin))

myeloid_distances_scoreA <- pivot_longer(myeloid_distances, cols = colnames(myeloid_distances)[str_detect(colnames(myeloid_distances), "activationCytokine")], names_to = "cytokines", values_to = "prop_cytokineExpressing")[,c(1,12,13)]
myeloid_distances_scoreA$cytokines <- sub("_expression_activationCytokine", "", myeloid_distances_scoreA$cytokines)

## Plot based
ggplot(myeloid_distances_scoreA, aes(cellsOI_distance_bin, prop_cytokineExpressing)) +
  #geom_point()+
  geom_smooth(method = "loess", aes(color = cytokines), se=FALSE)+
  #stat_cor(method="spearman")+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  ylab("Proportion of myeloid cells that express at least one cytokine")+
  xlab("Distance to the next T or NK cell expressing the given cytokine (5um bins)")


## Myeloid cell activation measured via activation score
myeloid_distances_scoreB<- pivot_longer(myeloid_distances, cols = colnames(myeloid_distances)[str_detect(colnames(myeloid_distances), "activationScore")], names_to = "cytokines", values_to = "activationScore")[,c(1,12,13)]
myeloid_distances_scoreB$cytokines <- sub("_expression_activationScore", "", myeloid_distances_scoreB$cytokines)

ggplot(myeloid_distances_scoreB, aes(cellsOI_distance_bin, activationScore)) +
  #geom_point()+
  geom_smooth(method = "lm", aes(color = cytokines), se=FALSE)+
  stat_cor(method="spearman", aes(color = cytokines))+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  ylab("Proportion of myeloid cells that express at least one cytokine")+
  xlab("Distance to the next T or NK cell expressing the given cytokine (5um bins)")

  # Remove CCL18 (is used to calculate the score --> biased)
  myeloid_distances_scoreB <- filter(myeloid_distances_scoreB, cytokines != "CCL18")
  
  # Calculate the mean over all cytokines other than CSF1
  scoreB <- pivot_wider(myeloid_distances_scoreB, values_from = "activationScore", names_from = "cytokines")
  scoreB_sub <- subset(scoreB, select=-c(CSF1, cellsOI_distance_bin))
  scoreB$meanActivationScore <- apply(scoreB_sub, MARGIN = 1, FUN = mean)
  
  ggplot(scoreB, aes(x = cellsOI_distance_bin)) +
  #geom_point()+
  geom_smooth(method = "lm", aes(y = CSF1), se=TRUE, color = "red")+
  stat_cor(method="spearman", aes(y = CSF1))+
  geom_smooth(method = "lm", aes(y = meanActivationScore), se=TRUE)+
  stat_cor(method="spearman", aes(y = meanActivationScore))+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  ylab("Proportion of myeloid cells that express at least one cytokine")+
  xlab("Distance to the next T or NK cell expressing the given cytokine (5um bins)")


```


### Use the expression of CD68, CD163, CD204 and CCL18 as a proxy for cell activation

```{r myeloid-activation-score-CSF1-distance, fig.width=5, fig.height=5}
# Re-normalize counts
norm_counts <- t(apply(assay(sce_myeloid, "counts"), 1, function(x)(x-min(x))/(quantile(x, 0.99)-min(x))))
norm_counts <- t(apply(norm_counts, 1, function(x) pmin(x, 1)))
assay(sce_myeloid, "normalized", withDimnames = FALSE) <- norm_counts

# For each cells, average across the normalized counts for CD68, CD163, CD204 and CCL18
count_matrix <- assay(sce_myeloid, "normalized")[c("CD68", "CD163", "CD204", "CCL18"),]
sce_myeloid$activation_score <- colMeans(count_matrix)

coldata_myeloid <- data.frame(colData(sce_myeloid))
myeloid_summarized <- coldata_myeloid %>% group_by(NKT_distance_bin) %>% summarize(mean_activation_score = mean(activation_score))
myeloid_summarized$distance <- seq(from = 7.5, to = 800, by = 5)

ggplot(coldata_myeloid, aes(NKT_distance_bin, activation_score)) +
  geom_boxplot()+
  coord_cartesian(ylim = c(0.11, 0.19))+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  ylab("Mean myeloid cell activation score")
  

ggplot(myeloid_summarized, aes(distance, mean_activation_score)) +
  geom_point()+
  geom_smooth(method = lm, color = "red", se=FALSE)+
  stat_cor(method="spearman")+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  ylab("Mean myeloid cell activation score")+
  xlab("Distance to the next NKT cells (5um bins)")

ggplot(coldata_myeloid, aes(NKT_distance, activation_score)) +
  geom_smooth(aes(group=sample, color = sample), method = loess)+
  #stat_cor(method="spearman")+
  #geom_smooth(method = loess, se = FALSE)+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  ylab("Mean myeloid cell activation score")+
  xlab("Distance to the next NKT cells (5um bins)")
```

# Save updated SCE object

```{r}
saveRDS(sce_RNA, "output/RNApanel/sce_RNApanel_cytokines_subtypes.rds")
```

# Compare proportions of CXCL13+ T cells to PD1+ T cells from HIER panel

```{r PD1-CXCL13-correlation, fig.width=5, fig.height=4}
sce_protein <- readRDS("output/ProteinPanel/sce_ProteinPanel_subtypes_all.rds")

# Proportion out of ALL cells
prop.CXCL13 <- prop.table(table(sce_RNA$ImageNumber, sce_RNA$subtype))[,c("T_CD4_CXCL13", "T_CD8_CXCL13")]
prop.PD1 <- prop.table(table(sce_protein$ImageNumber, sce_protein$subtype))[,c("CD4_PD1", "CD8_PD1")]
prop <- cbind(prop.CXCL13, prop.PD1)

# Add ROI info
ROI_info <- read.csv("data/ROI_info_RNAscope.csv")
prop <- cbind(prop, ROI_info)
prop$sample <- factor(prop$sample, levels = levels(sce_RNA$sample))

# CD4+ T cells
ggplot(prop, aes(T_CD4_CXCL13, CD4_PD1)) +
  geom_point(aes(color=sample))+
  geom_smooth(method = lm, color = "black", se=FALSE)+
  stat_cor(method="spearman")+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())

# CD8+ T cells
ggplot(prop, aes(T_CD8_CXCL13, CD8_PD1)) +
  geom_point(aes(color=sample))+
  geom_smooth(method = lm, color = "black", se=FALSE)+
  stat_cor(method="spearman")+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())

# CD4+ T cells
ggplot(prop, aes(T_CD4_CXCL13, CD4_PD1)) +
  geom_point(aes(color=sample))+
  geom_smooth(method = lm, color = "black", se=FALSE)+
  stat_cor(method="spearman")+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())

# All T cells
ggplot(prop, aes(T_CD8_CXCL13+T_CD4_CXCL13, CD8_PD1+CD4_PD1)) +
  geom_point(aes(color=sample))+
  geom_smooth(method = lm, color = "black", se=FALSE)+
  stat_cor(method="spearman")+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())+
  scale_color_manual(values = sce_TNK@metadata$colour_vectors$sample)
```
