==> analysis/11_RNAscope_cytokine_communities.Rmd <==
---
title: "11_RNAscope_cytokine_communities"
author: "SandraTietscher"
date: "2020-11-03"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

# Introduction

This script uses Tobi's functions to detect interaction networks of a given celltype (here: cytokine producing cells) and defines these networks as clusters. Once a cluster is defined, an algorithm screens the neighbourhood of those clusters to identify cells within/surrounding a cluster. These cells are defined as the community of a cluster. 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message= FALSE)
```

# Preparations

# Load packages, helper functions and data

```{r message=FALSE, warning=FALSE}
sapply(list.files("code/helper_functions", full.names = TRUE), source)
library(SingleCellExperiment)
library(ggplot2)
library(scater)
library(igraph)
library(reshape2)
library(cowplot)
library(ggridges)
library(tidyverse)
library(viridis)
library(dplyr)
library(sp)
library(sf)
library(RANN)
library(concaveman)
library(ggpubr)
library(ComplexHeatmap)
library(RColorBrewer)
#library(V8)

sce = readRDS(file = "output/RNApanel/sce_RNApanel_cytokines_subtypes.rds")

cytokine_channels = c("CCL4", "CCL18", "CSF1", "CXCL10", "CXCL13_mRNA", "CCL2", "CCL22", "CXCL9", "CCL17", "CCL5")
TIG2 <- c("TBB111", "TBB129", "TBB165", "TBB171", "TBB184", "TBB338")
ROI_info <- read.csv("data/ROI_info_RNAscope.csv")
ROI_info$ImageNumber <- 1:77 
```

# Clusters/Communities - General

## Find clusters of cytokine-producing cells and define communities

A cluster of cytokine-producing cells consists of cells (min. 10) that have a max distance of 20µm to the next cytokine producing cells. 
A community is defined by cells which surround (max 25µm) the border of a cluster (definded by the convex hull of the cluster).  

```{r Cluster and Community Detection cytokines general}
# average cell diameter
sqrt(mean(sce$Area)/pi)
sqrt(median(sce$Area)/pi)
quantile(sqrt(sce$Area/pi))

# find cytokine clusters
sce <- findClusters(sce, sce[,colData(sce)$cytokine == TRUE]$cellID , 
                    'cellID', 
                    'Center_X', 'Center_Y', 
                    'ImageNumber', 
                    distance = 20, 
                    min_clust_size = 10,
                    output_colname = "cytokine_cluster")

# number of cytokine clusters
length(unique(sce$cytokine_cluster))

# define cells within/surrounding a cluster of cytokine producing cells
sce <- findCommunity(sce, 
                     'cellID', 
                     'Center_X', 'Center_Y', 
                     'ImageNumber', 
                     'cytokine_cluster', 
                     distance = 25,
                     output_colname = "cytokine_community")

# number of cytokine communities
length(unique(sce$cytokine_community))
```

# Detect Clusters and Commmunities for Single cytokines

```{r cluster-detection-single}
start = Sys.time()
cur_sce <- data.frame(colData(sce))

# loop through all cytokines
for(i in cytokine_channels) {

  # find clusters
  sce <- findClusters(input_sce = sce, 
                    IDs_of_interest = cur_sce[cur_sce[,names(cur_sce) == paste0(i, "_expression")] == 1,]$cellID, 
                    'cellID', 
                    'Center_X', 'Center_Y', 
                    'ImageNumber', 
                    distance = 25, 
                    min_clust_size = 3,
                    output_colname = paste(tolower(i), "only_clust", sep = ""))

  
  # define cells within/surrounding a cluster of cytokine producing cells
  sce <- findCommunity(sce, 
                     'cellID', 
                     'Center_X', 
                     'Center_Y', 
                     'ImageNumber', 
                     paste(tolower(i), "only_clust", sep = ""), 
                     distance = 30,
                     output_colname = paste(tolower(i), "only_comm", sep = ""))
} 
end = Sys.time()
print(end-start)
```

# Save SCE object

```{r save-sce}
saveRDS(sce, file = "output/RNApanel/sce_RNApanel_cytokines_communities.rds")
```

# Cluster size distribution

```{r cluster-sizes}
#For general cytokine-expressing cluster
cytokine_cluster <- as.data.frame(table(sce$cytokine_cluster)[-1])
ggplot(cytokine_cluster, aes(x=Freq))+
  geom_histogram(binwidth = 10)

#For individual single-cytokine clusters
cur_sce <- data.frame(colData(sce))
clust_colnames <- colnames(cur_sce)[grep("only_clust", colnames(cur_sce))]

p.list <- list()
for (i in clust_colnames) {
  cluster <- as.data.frame(table(cur_sce[,i])[-1])
  p.list[[i]] <- ggplot(cluster, aes(x=Freq))+ geom_histogram(binwidth = 1) + ggtitle(i) + ylab("Cluster size")
  
}
plot_grid(plotlist = p.list, ncol = 4)
```

# Community size distribution

```{r comminity-sizes}
#For general cytokine-expressing cluster
cytokine_community <- as.data.frame(table(sce$cytokine_community)[-1])
ggplot(cytokine_community, aes(x=Freq))+
  geom_histogram(binwidth = 10)

#For individual single-cytokine clusters
comm_colnames <- colnames(cur_sce)[grep("only_comm$", colnames(cur_sce))]

p.list <- list()
for (i in comm_colnames) {
  comm <- as.data.frame(table(cur_sce[,i])[-1])
  p.list[[i]] <- ggplot(comm, aes(x=Freq))+ geom_histogram(binwidth = 1) + ggtitle(i) + ylab("Community size")
  
}
plot_grid(plotlist = p.list, ncol = 4)
```

# Cytomapper: cytokine cluster plots

```{r cytomapper-cluster, message = FALSE}
library(cytomapper)

# Load mask files
path.to.images <- "data/masks/RNApanel_masks"
all_masks <- loadImages(path.to.images, pattern = "_mask.tiff")

# Add image number
mcols(all_masks)$ImageNumber <- c(1:77)

# Scale images
all_masks <- scaleImages(all_masks, 2^16-1)
head(unique(as.numeric(all_masks[[1]])))

cur_img <- getImages(all_masks, "20200817_ST_BCexh_R_TBB165_s0_p3_r1_a1_ac_ilastik_s2_Probabilities_mask")

# Plot cells
plotCells(cur_img, object = sce,
            img_id = "ImageNumber", cell_id = "CellNumber",
            colour_by = "ccl22only_clust")

plotCells(cur_img, object = sce,
            img_id = "ImageNumber", cell_id = "CellNumber",
            colour_by = c("CCL22"), exprs_values = "exprs")

```

# Cell type proportions

```{r single-vs-cluster-all, fig.width=7, fig.height=5}
sce_cyt <- sce[,which(sce$cytokine == 1)]
sce_cyt <- sce_cyt[,which(sce_cyt$TLS == "no")]
sce_cyt$single <- ifelse(sce_cyt$cytokine_cluster == 0, "single", "cluster")

tab_sample <- as.data.frame(prop.table(table(sce_cyt$sample, sce_cyt$single), margin = 1))
tab_sample <- filter(tab_sample, Var2 == "cluster")
tab_sample$TIG <- ifelse(tab_sample$Var1 %in% TIG2, "TIG2", "TIG3")
colnames(tab_sample)[1:2] <- c("sample", "cluster_single")

ggplot(tab_sample, aes(TIG, Freq, color = TIG))+
  geom_boxplot(outlier.shape = NA)+
  geom_jitter(width = 0.1)+
  stat_compare_means(method = "wilcox.test", aes(label=..p.format..), label.x.npc = 0.5, label.y.npc = 0.95)+
  theme(axis.title.x=element_blank(),panel.background = element_blank(), axis.text.x = element_text(angle = 60, hjust = 1), legend.title = element_blank(), panel.border = element_rect(color = "black", fill = NA), strip.background = element_blank())+
  ylab("Proportion of cytokine-expressing cells that are part of a cytokine cluster")
```

```{r single-vs-cluster, fig.width=7, fig.height=5}
sce_cyt <- sce[,which(sce$cytokine == 1)]
sce_cyt <- sce_cyt[,which(sce_cyt$TLS == "no")]
sce_cyt$single <- ifelse(sce_cyt$cytokine_cluster == 0, "single", "cluster")

tab_celltype <- data.frame(matrix(ncol=4, nrow = 0))

for (i in unique(sce$sample)){
  sce_i <- sce_cyt[,which(sce_cyt$sample == i)]
  tab_celltype_i <- as.data.frame(prop.table(table(sce_i$celltype, sce_i$single), margin = 1))
  tab_celltype_i$sample <- i
  tab_celltype <- rbind(tab_celltype, tab_celltype_i)
}

tab_celltype <- filter(tab_celltype, Var2 == "cluster")
tab_celltype$TIG <- ifelse(tab_celltype$sample %in% TIG2, "TIG2", "TIG3")
colnames(tab_celltype)[1:2] <- c("celltype", "cluster_single")

ggplot(tab_celltype, aes(TIG, Freq, color = TIG))+
  geom_boxplot(outlier.shape = NA)+
  geom_jitter(width = 0.1)+
  facet_wrap(~celltype, ncol = 6)+
  stat_compare_means(method = "wilcox.test", aes(label=..p.format..), label.x.npc = 0.5, label.y.npc = 0.95)+
  theme(axis.title.x=element_blank(),panel.background = element_blank(), axis.text.x = element_text(angle = 60, hjust = 1), legend.title = element_blank(), panel.border = element_rect(color = "black", fill = NA), strip.background = element_blank())+
  ylab("Proportion of cytokine-expressing cells that are part of a cytokine cluster")
```

```{r single-vs-cluster-tumor-immune, fig.width=5, fig.height=4}
sce_cyt$cell_class <- ifelse(sce_cyt$celltype == "tumor", "tumor", ifelse(sce_cyt$celltype %in% c("endothelial", "stromal_SMA+", "stromal_undefined"), "stromal", "immune"))

tab_cellclass <- as.data.frame(prop.table(table(sce_cyt$cell_class, sce_cyt$single), margin = 1))
colnames(tab_cellclass) <- c("cell_class", "single_cluster", "proportion")

ggplot(tab_cellclass, aes(cell_class, proportion, fill = single_cluster))+
  geom_bar(stat = "identity")+
  scale_fill_manual(values = c("#E69F00", "#006600"))+
  theme(axis.title.x=element_blank(),panel.background = element_blank(), axis.text.x = element_text(angle = 60, hjust = 1), legend.title = element_blank(), panel.border = element_rect(color = "black", fill = NA))
```

```{r single-vs-cluster-TLS, fig.width=7, fig.height=5}
sce_cyt <- sce[,which(sce$cytokine == 1)]
sce_cyt$single <- ifelse(sce_cyt$cytokine_cluster == 0, "single", "cluster")

tab_celltype <- data.frame(matrix(ncol=4, nrow = 0))

for (i in unique(sce$ImageNumber)){
  sce_i <- sce_cyt[,which(sce_cyt$ImageNumber == i)]
  tab_celltype_i <- as.data.frame(prop.table(table(sce_i$celltype, sce_i$single), margin = 1))
  tab_celltype_i$ImageNumber <- i
  tab_celltype <- rbind(tab_celltype, tab_celltype_i)
}

tab_celltype <- filter(tab_celltype, Var2 == "single")

colnames(tab_celltype)[1:2] <- c("celltype", "cluster_single")
tab_celltype$inCluster_freq <- 1 - tab_celltype$Freq

# Add TLS info
tab_celltype$TLS <- ROI_info$TLS[match(tab_celltype$ImageNumber, ROI_info$ImageNumber)]
my_comparisons <- list(c("no", "pre"), c("no", "yes"), c("pre", "yes"))

ggplot(tab_celltype, aes(TLS, inCluster_freq))+
  geom_boxplot()+
  #geom_jitter(width = 0.1)+
  facet_wrap(~celltype, ncol = 6)+
  theme(axis.title.x=element_blank(),panel.background = element_blank(), axis.text.x = element_text(angle = 60, hjust = 1), legend.title = element_blank(), panel.border = element_rect(color = "black", fill = NA), strip.background = element_blank())+
  ylab("Proportion of cytokine-expressing cells that are part of a cytokine cluster")+
  stat_compare_means(comparisons = my_comparisons, method = "wilcox.test", aes(label=..p.signif..))
```


## Cell type proportions in each cluster

```{r cytokine-cluster-celltypes}
# For individual single-cytokine clusters

celltype_prop <- data.frame(celltype = names(prop.table(table(cur_sce[,"celltype"]))))
celltype_prop[,"all"] <- as.numeric(prop.table(table(cur_sce[,"celltype"])))

for (i in clust_colnames) {
  cur_sce_i <- cur_sce[which(cur_sce[,i] != 0),]
  prop_i <- as.data.frame(prop.table(table(cur_sce_i[,"celltype"])))
  colnames(prop_i) <- c("celltype", i)
  celltype_prop <- merge(celltype_prop, prop_i, by="celltype", all.x = TRUE)
}

celltype_prop[is.na(celltype_prop)] = 0
celltype_prop_simple <- pivot_longer(celltype_prop, 2:12, names_to = "cluster", values_to = "celltype_proportion")

# Plot
ggplot(celltype_prop_simple, aes(cluster, celltype_proportion, fill = celltype))+
  geom_bar(stat = "identity")+
    theme(axis.title.x=element_blank(),panel.background = element_blank(), axis.text.x = element_text(angle = 60, hjust = 1))+
  ggtitle("Cell type proportions per cytokine cluster")
```

## Cell type proportions in communities

```{r community-celltypes-all}
# For individual single-cytokine communities: ALL CELLS IN THE COMMUNITY

celltype_comm <- data.frame(celltype = names(prop.table(table(cur_sce[,"celltype"]))))
celltype_comm[,"all"] <- as.numeric(prop.table(table(cur_sce[,"celltype"])))

for (i in comm_colnames) {
  cur_sce_i <- cur_sce[which(cur_sce[,i] != 0),]
  prop_i <- as.data.frame(prop.table(table(cur_sce_i[,"celltype"])))
  colnames(prop_i) <- c("celltype", i)
  celltype_comm <- merge(celltype_comm, prop_i, by="celltype", all.x = TRUE)
}

celltype_comm[is.na(celltype_comm)] = 0
celltype_comm <- pivot_longer(celltype_comm, 2:12, names_to = "cluster", values_to = "celltype_proportion")

# Plot
ggplot(celltype_comm, aes(cluster, celltype_proportion, fill = celltype))+
  geom_bar(stat = "identity")+
    theme(axis.title.x=element_blank(),panel.background = element_blank(), axis.text.x = element_text(angle = 60, hjust = 1))+
  ggtitle("Cell type proportions per cytokine community")
```

```{r community-enrichment-all}
cur_sce$celltype <- factor(cur_sce$celltype, levels = unique(cur_sce$celltype))

#Calculate enrichment individually for each community (to prevent bias from large communities)
enrichment_comm <- data.frame(matrix(ncol=4, nrow = 0))

for (i in cytokine_channels) {
  comm_name_i <- tolower(paste0(i,"only_comm"))
  cur_sce_i <- cur_sce[which(cur_sce[,comm_name_i] != 0),]
   
    for(j in unique(cur_sce_i[,comm_name_i])) {
      sce_sub_j <- cur_sce_i[cur_sce_i[,comm_name_i] == j,]
      prop_j <- prop.table(table(sce_sub_j$celltype))
      prop_all <- prop.table(table(cur_sce$celltype))[names(prop_j)]
      enrichment_j <-  prop_j/prop_all 
      if(length(prop_j > 0)) {
            j_df <- data.frame(comm = j,
                        cytokine = i,
                        enrichment = enrichment_j)
      enrichment_comm <- rbind(enrichment_comm, j_df)
      }
  }
}

colnames(enrichment_comm)[3:4] = c("celltype", "enrichment")

ggplot(enrichment_comm, aes(cytokine, enrichment, fill = cytokine))+
  geom_boxplot()+
  facet_wrap(~celltype, scales = "free", ncol = 4)+
  theme(panel.background = element_blank(),
        axis.text.x = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  geom_hline(yintercept = 1, linetype = "dashed")
```

### For individual single-cytokine communities: ONLY CELLS IN THE COMMUNITY THAT ARE NOT expressing the given cytokine

```{r community-celltypes-nonexpressing, fig.width=7, fig.height=5.7}

celltype_commonly <- data.frame(celltype = names(prop.table(table(cur_sce[,"celltype"]))))
celltype_commonly[,"all"] <- as.numeric(prop.table(table(cur_sce[,"celltype"])))

for (i in cytokine_channels) {
  cur_sce_i <- cur_sce[which(cur_sce[,str_detect(colnames(cur_sce), regex(paste0(i,"only_comm"), ignore_case = TRUE))] != 0 & cur_sce[,str_detect(colnames(cur_sce), regex(paste0(i,"_expression"), ignore_case = TRUE))] == 0),]
  prop_i <- as.data.frame(prop.table(table(cur_sce_i[,"celltype"])))
  colnames(prop_i) <- c("celltype", i)
  celltype_commonly <- merge(celltype_commonly, prop_i, by="celltype", all.x = TRUE)
}

celltype_commonly[is.na(celltype_commonly)] = 0
celltype_commonly_long <- pivot_longer(celltype_commonly, 2:12, names_to = "cluster", values_to = "celltype_proportion")

# Plot
ggplot(celltype_commonly_long, aes(cluster, celltype_proportion, fill = celltype))+
  geom_bar(stat = "identity")+
    theme(axis.title.x=element_blank(),panel.background = element_blank(), axis.text.x = element_text(angle = 60, hjust = 1))+
  ggtitle("Cells that are ONLY part of the respective cytokine community (not the cluster)")

# Enrichment compared to overall proportion
celltype_enrichment <- celltype_commonly
for (i in 3:12) {
  celltype_enrichment[i] <- celltype_commonly[i]/celltype_commonly[2]
}
celltype_enrichment_long <- pivot_longer(celltype_enrichment, 3:12, names_to = "cluster", values_to = "enrichment")

# Plot
ggplot(celltype_enrichment_long, aes(cluster, enrichment, fill = cluster))+
  geom_bar(stat="identity")+
  facet_wrap(~celltype, scales = "free")+
  theme(panel.background = element_blank(),
        axis.text.x = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  geom_hline(yintercept = 1, linetype = "dashed")
```

#### Calculate significance using Fisher's exact test

```{r community-enrichment-nonexpressing-fisher}

cur_sce <- data.frame(colData(sce))

## Calculate enrichment individually for each image
cur_sce$celltype <- factor(cur_sce$celltype, levels = unique(cur_sce$celltype))

enrich_matrix <- matrix(data = 0, ncol=length(cytokine_channels), nrow = length(unique(cur_sce$celltype)))
colnames(enrich_matrix) <- cytokine_channels
rownames(enrich_matrix) <- unique(cur_sce$celltype)

#Initiate vector to count images that have a community of the respective cytokine
imagenr <- rep(0, length(cytokine_channels))
names(imagenr) <- cytokine_channels

#Initiate matrix to count images that have a community of the respective cytokine AND contain the respective celltype
imagenr_celltype <- enrich_matrix

# Loop over cytokine channels
for (i in cytokine_channels) {
  comm_name_i <- tolower(paste0(i,"only_comm"))
  exp_name_i <- colnames(cur_sce)[str_detect(colnames(cur_sce), string = regex(paste0(i,"_expression"), ignore_case = TRUE))]
  
  # Define cells that are within the community but do not themselves express the cytokine
  cur_sce_i <- cur_sce[which(cur_sce[,comm_name_i] != 0 & cur_sce[,exp_name_i] == 0),]
    
  # Exclude cells that express the given cytokine from full dataset (they would otherwise be counted as "outside" the community)
  cur_sce_wo_patch <- cur_sce[which(cur_sce[,exp_name_i] == 0),]
   
  #Loop over images
    for(j in unique(cur_sce_i$ImageNumber)) {
      sce_sub_ji <- cur_sce_i[cur_sce_i[,"ImageNumber"] == j,]
      
      # Only proceed if the community i is present in image j
      if (ncol(sce_sub_ji) > 0) {
        imagenr[i] <- imagenr[i] + 1
        
        #Loop over cell types
        for(k in unique(cur_sce_i$celltype)) {
          count_kimg <- table(cur_sce_wo_patch$celltype, cur_sce_wo_patch$ImageNumber)[k, j]  # Number of this cell in the image
          
          # Only proceed if the celltype k is present in image j (otherwise we will have to divide by 0 which leads to NA's)
          if (count_kimg > 0) {
            imagenr_celltype[k, i] <- imagenr_celltype[k,i] + 1
            count_kj <- table(sce_sub_ji$celltype)[k]   # Number of this cell in community k
            count_j <- nrow(sce_sub_ji)                # Total number of cells in community k
            count_kimg <- table(cur_sce_wo_patch$celltype, cur_sce_wo_patch$ImageNumber)[k, j]  # Number of this cell in the image
            count_img <- table(cur_sce_wo_patch$ImageNumber)[j]  # Number cells in the image
            
            count.matrix <- matrix(c(count_kj, count_j, count_kimg, count_img), nrow=2, ncol=2)
            p <- fisher.test(count.matrix)[["p.value"]]
            enrich_depl <- ifelse(count_kj/count_kimg > count_j/count_img, 1, -1)
            enrich_value <- ifelse(p < 0.01, enrich_depl, 0)
            
            enrich_matrix[k, i] <- enrich_matrix[k, i] + enrich_value
          }
        }
      }
  }
}

# Divide enrichment matrix through image number with the respective cytokine community
enrich_corr <- apply(enrich_matrix, MARGIN = 1, function(x) x/imagenr)
enrich_corr[is.na(enrich_corr)] <- 0
enrich_corr <- enrich_corr*100

cols = rev(brewer.pal(11,'Spectral'))
cmap = colorRampPalette(cols)
Heatmap(enrich_corr, column_title = "% of images with significant enrichment/depletion", row_title = "Cytokine milieu", row_title_side = "right", heatmap_legend_param = list(title = ""), col = cmap(75))

# Divide enrichment matrix through image number (with the respective cytokine community), in which the respective celltype is present --> result is almost identical
enrich_corr_celltype <- t(enrich_matrix/imagenr_celltype)
Heatmap(enrich_corr_celltype, column_title = "% of images with significant enrichment/depletion", row_title = "Cytokine milieu", row_title_side = "right", heatmap_legend_param = list(title = ""), col = cmap(75))

```

Use cell subtypes instead of cell types

```{r community-enrichment-chisquare-subtypes}

## Calculate enrichment individually for each image

cur_sce$subtype <- factor(cur_sce$subtype, levels = unique(cur_sce$subtype))

enrich_matrix <- matrix(data = 0, ncol=length(cytokine_channels), nrow = length(unique(cur_sce$subtype)))
colnames(enrich_matrix) <- cytokine_channels
rownames(enrich_matrix) <- unique(cur_sce$subtype)

#Initiate vector to count images that have a community of the respective cytokine
imagenr <- rep(0, length(cytokine_channels))
names(imagenr) <- cytokine_channels

#Initiate matrix to count images that have a community of the respective cytokine AND contain the respective subtype
imagenr_subtype <- enrich_matrix

# Loop over cytokine channels
for (i in cytokine_channels) {
  comm_name_i <- tolower(paste0(i,"only_comm"))
  cur_sce_i <- cur_sce[which(cur_sce[,comm_name_i] != 0 & cur_sce[,str_detect(colnames(cur_sce), string = regex(paste0(i,"_expression"), ignore_case = TRUE))] == 0),]
    #cur_sce_i <- cur_sce[which(cur_sce[,comm_name_i] != 0),]
   
  #Loop over images
    for(j in unique(cur_sce_i$ImageNumber)) {
      sce_sub_ji <- cur_sce_i[cur_sce_i[,"ImageNumber"] == j,]
      
      # Only proceed if the community i is present in image j
      if (ncol(sce_sub_ji) > 0) {
        imagenr[i] <- imagenr[i] + 1
        
        #Loop over cell types
        for(k in unique(cur_sce_i$subtype)) {
          count_kimg <- table(cur_sce$subtype, cur_sce$ImageNumber)[k, j]  # Number of this cell in the image
          
          # Only proceed if the subtype k is present in image j (otherwise we will have to divide by 0 which leads to NA's)
          if (count_kimg > 0) {
            imagenr_subtype[k, i] <- imagenr_subtype[k,i] + 1
            count_kj <- table(sce_sub_ji$subtype)[k]   # Number of this cell in community k
            count_j <- nrow(sce_sub_ji)                # Total number of cells in community k
            count_kimg <- table(cur_sce$subtype, cur_sce$ImageNumber)[k, j]  # Number of this cell in the image
            count_img <- table(cur_sce$ImageNumber)[j]  # Number cells in the image
            
            chisq.matrix <- matrix(c(count_kj, count_j, count_kimg, count_img), nrow=2, ncol=2)
            p <- chisq.test(chisq.matrix)[["p.value"]]
            enrich_depl <- ifelse(count_kj/count_kimg > count_j/count_img, 1, -1)
            enrich_value <- ifelse(p < 0.01, enrich_depl, 0)
            
            enrich_matrix[k, i] <- enrich_matrix[k, i] + enrich_value
          }
        }
      }
  }
}

# Divide enrichment matrix through image number with the respective cytokine community
enrich_corr <- apply(enrich_matrix, MARGIN = 1, function(x) x/imagenr)
enrich_corr[is.na(enrich_corr)] <- 0
enrich_corr <- enrich_corr*100

cols = rev(brewer.pal(11,'Spectral'))
cmap = colorRampPalette(cols)
Heatmap(enrich_corr, column_title = "% of images with significant enrichment/depletion", row_title = "Cytokine milieu", row_title_side = "right", heatmap_legend_param = list(title = ""), col = cmap(75))

# Divide enrichment matrix through image number (with the respective cytokine community), in which the respective subtype is present --> result is almost identical
enrich_corr_subtype <- t(enrich_matrix/imagenr_subtype)
Heatmap(enrich_corr_subtype, na_col = "gray")


```

ONLY FOR TIG2

```{r community-enrichment-nonexpressing-TIG2}

## Calculate enrichment individually for each community (to prevent bias from large communities)

enrichment_comm <- data.frame(matrix(ncol=4, nrow = 0))
cur_sce_TIG2 <- cur_sce[which(cur_sce$TIG == "TIG2"),]

for (i in cytokine_channels) {
  comm_name_i <- tolower(paste0(i,"only_comm"))
  cur_sce_i <- cur_sce_TIG2[which(cur_sce_TIG2[,comm_name_i] != 0 & cur_sce_TIG2[,str_detect(colnames(cur_sce_TIG2), regex(paste0(i,"_expression"), ignore_case = TRUE))] == 0),]
   
    for(j in unique(cur_sce_i[,comm_name_i])) {
      sce_sub_j <- cur_sce_i[cur_sce_i[,comm_name_i] == j,]
      prop_j <- prop.table(table(sce_sub_j$celltype))
      prop_all <- prop.table(table(cur_sce_TIG2$celltype))[names(prop_j)]
      enrichment_j <-  prop_j/prop_all 
      if(length(prop_j > 0)) {
            j_df <- data.frame(comm = j,
                        cytokine = i,
                        enrichment = enrichment_j)
      enrichment_comm <- rbind(enrichment_comm, j_df)
      }
  }
}

colnames(enrichment_comm)[3:4] = c("celltype", "enrichment")

ggplot(enrichment_comm, aes(cytokine, enrichment, fill = cytokine))+
  geom_boxplot(outlier.alpha = 0.2)+
  facet_wrap(~celltype, scales = "free", ncol = 6)+
  theme(panel.background = element_blank(),
        axis.text.x = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  geom_hline(yintercept = 1, linetype = "dashed")
```

ONLY FOR TIG3

```{r community-enrichment-nonexpressing-TIG3}

## Calculate enrichment individually for each community (to prevent bias from large communities)

enrichment_comm <- data.frame(matrix(ncol=4, nrow = 0))
cur_sce_TIG3 <- cur_sce[which(cur_sce$TIG == "TIG3"),]

for (i in cytokine_channels) {
  comm_name_i <- tolower(paste0(i,"only_comm"))
  cur_sce_i <- cur_sce_TIG3[which(cur_sce_TIG3[,comm_name_i] != 0 & cur_sce_TIG3[,str_detect(colnames(cur_sce_TIG3), regex(paste0(i,"_expression"), ignore_case = TRUE))] == 0),]
   
    for(j in unique(cur_sce_i[,comm_name_i])) {
      sce_sub_j <- cur_sce_i[cur_sce_i[,comm_name_i] == j,]
      prop_j <- prop.table(table(sce_sub_j$celltype))
      prop_all <- prop.table(table(cur_sce_TIG3$celltype))[names(prop_j)]
      enrichment_j <-  prop_j/prop_all 
      if(length(prop_j > 0)) {
            j_df <- data.frame(comm = j,
                        cytokine = i,
                        enrichment = enrichment_j)
      enrichment_comm <- rbind(enrichment_comm, j_df)
      }
  }
}

colnames(enrichment_comm)[3:4] = c("celltype", "enrichment")

ggplot(enrichment_comm, aes(cytokine, enrichment, fill = cytokine))+
  geom_boxplot(outlier.alpha = 0.2)+
  facet_wrap(~celltype, scales = "free", ncol = 6)+
  theme(panel.background = element_blank(),
        axis.text.x = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  geom_hline(yintercept = 1, linetype = "dashed")
```

## Compare cell type proportions in clusters and communities

```{r cluster-community-celltype-comparison, fig.width= 15, fig.height=8}
# Compare
celltype_prop_simple$cluster <- celltype_commonly_long$cluster
celltype_prop_simple$type <- "cluster_only"
celltype_commonly_long$type <- "community_only"
celltype_comb <- rbind(celltype_prop_simple, celltype_commonly_long)
celltype_comb <- celltype_comb %>% filter(cluster != "all")

p.list <- list()
for (i in unique(celltype_comb$cluster)) {
celltype_comb_i <- filter(celltype_comb, cluster == i)
p.list[[i]] <- ggplot(celltype_comb_i, aes(type, celltype_proportion, fill = celltype))+
  geom_bar(stat = "identity", width = 0.6)+
  geom_segment(data = tidyr::spread(celltype_comb_i, type, celltype_proportion),
               colour = "grey",
               aes(x = 1 + 0.6/2,
                   xend = 2 - 0.6/2,
                   y = 1 - cumsum(cluster_only),
                   yend = 1 - cumsum(community_only))) +
    theme(axis.title.x=element_blank(),panel.background = element_blank(), axis.text.x = element_text(angle = 60, hjust = 1), legend.position = "none")+
  ggtitle(i)
}

plot_grid(plotlist = p.list, ncol = 5)
```

# CSF1 environments - myeloid cell activation

Check whether myeloid cells in CSF1 environments show a higher degree of activation than in other environments/outside these environments.

### Use cytokine expression as a proxy for myeloid cell activation
```{r CSF1-myeloid-cytokine, fig.width=4, fig.height=3}
sce_myeloid <- sce[,which(sce$celltype == "myeloid")]
perc_total <- data.frame(total = prop.table(table(sce_myeloid$cytokine))[2])
coldata_myeloid <- data.frame(colData(sce_myeloid))

# only select cells that are in the community, but do not themselves express the cytokine
for (i in cytokine_channels) {
  comm_name_i <- tolower(paste0(i,"only_comm"))
  cur_sce_i <- coldata_myeloid[which(coldata_myeloid[,comm_name_i] != 0 & coldata_myeloid[,str_detect(colnames(coldata_myeloid), string = regex(paste0(i,"_expression"), ignore_case = TRUE))] == 0),]
  perc_i <- prop.table(table(cur_sce_i$cytokine))[2]
  perc_total[,i] <- perc_i
}

perc_total <- pivot_longer(perc_total, 1:11, names_to = "environment", values_to = "percent_cytokine_expressing")

ggplot(perc_total, aes(environment, percent_cytokine_expressing, fill = environment))+
  geom_bar(stat="identity")+
  theme(panel.background = element_blank(),
        axis.text.x = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ggtitle("Cytokine-expressing myeloid percentage\nin the respective communities")

```

### Use the expression of CD68, CD163, CD204 and CCL18 as a proxy for cell activation

```{r CSF1-myeloid-activation-score}
# Re-normalize counts
norm_counts <- t(apply(assay(sce_myeloid, "counts"), 1, function(x)(x-min(x))/(quantile(x, 0.99)-min(x))))
norm_counts <- t(apply(norm_counts, 1, function(x) pmin(x, 1)))
assay(sce_myeloid, "normalized", withDimnames = FALSE) <- norm_counts

# For each cells, average across the normalized counts for CD68, CD163, CD204 and CCL18
count_matrix <- assay(sce_myeloid, "normalized")[c("CD68", "CD163", "CD204", "CCL18"),]
sce_myeloid$activation_score <- colMeans(count_matrix)

activation_mean <- data.frame(total = mean(sce_myeloid$activation_score))
coldata_myeloid <- data.frame(colData(sce_myeloid))

# only select cells that are in the community, but NOT in the corresponding cluster
for (i in cytokine_channels) {
 comm_name_i <- tolower(paste0(i,"only_comm"))
  cur_sce_i <- coldata_myeloid[which(coldata_myeloid[,comm_name_i] != 0 & coldata_myeloid[,str_detect(colnames(coldata_myeloid), string = regex(paste0(i,"_expression"), ignore_case = TRUE))] == 0),]
  activation_mean[,i] <- mean(cur_sce_i$activation_score)
}

activation_mean <- pivot_longer(activation_mean, 1:11, names_to = "environment", values_to = "mean_activation_score")

ggplot(activation_mean, aes(environment, mean_activation_score, fill = environment))+
  geom_bar(stat="identity")+
  theme(panel.background = element_blank(),
        axis.text.x = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())
```


# Cytokine environments -> apply to matched ROIs

Define "cytokine environments" by extending the respective cytokine communities by 10um (concave hull), thus allowing for slight mismatches between the different ROIs.   
Extract coordinates for these environments and lay them over the matched ROIs from the protein panel.   
Output: New protein panel SCE with added ColData about which cells belong to which cytokine environments.

The goal is to investigate association of cytokine expression with celltypes that can only be defined in the Protein Panel (e.g. regulatory T cells).
  
```{r cytokine-environment-proteinpanel}
# read in SCE from protein panel
sce_protein <- readRDS("output/ProteinPanel/sce_ProteinPanel_subtypes_all.rds")

# ColData column names for cytokine communities
comm_colnames <- c(colnames(colData(sce))[grep("only_comm", colnames(colData(sce)))],"cytokine_community")

# Set distance parameter
distance = 10

for (k in comm_colnames) {
  
input_df_RNA <- data.frame(cellID = sce$cellID,
                       X = sce$Center_X,
                       Y = sce$Center_Y,
                       ImageNumber = sce$ImageNumber,
                       cluster = colData(sce)[,k])

input_df_protein <- data.frame(cellID = sce_protein$cellID,
                       X = sce_protein$Center_X,
                       Y = sce_protein$Center_Y,
                       ImageNumber = sce_protein$ImageNumber)

# Create empty dataframe to collect all cells in the respective cytokine environment (for protein panel)
k_env_cells <- data.frame(matrix(ncol=ncol(input_df_protein), nrow = 0))
colnames(k_env_cells) = colnames(input_df_protein)

# Create empty dataframe to collect all cells in the respective cytokine environment (for protein panel)
k_env_cells_RNA <- data.frame(matrix(ncol=ncol(input_df_RNA), nrow = 0))
colnames(k_env_cells_RNA) = colnames(input_df_RNA)

  for (j in unique(input_df_RNA$ImageNumber)) {
    
    # create multipoint object with all cells from one image (for RNA panel)
      cell_coord = as.matrix(input_df_RNA[input_df_RNA$ImageNumber == j,2:3])
      rownames(cell_coord) = input_df_RNA[input_df_RNA$ImageNumber == j,1]
      cells = sf::st_multipoint(cell_coord)
      cells_sfc = sf::st_cast(sf::st_sfc(cells), "POINT")
      
    # create multipoint object with all cells from one image (for Protein panel)
      cell_coord_protein = as.matrix(input_df_protein[input_df_protein$ImageNumber == j,2:3])
      rownames(cell_coord_protein) = input_df_protein[input_df_protein$ImageNumber == j,1]
      cells_protein = sf::st_multipoint(cell_coord_protein)
      cells_sfc_protein = sf::st_cast(sf::st_sfc(cells_protein), "POINT")
      
    # create data.frame for all environment cells for the respective image (for protein panel)
      environment_cells = data.frame(matrix(ncol=ncol(input_df_protein), nrow = 0))
      colnames(environment_cells) = colnames(input_df_protein)
      
    # create data.frame for all environment cells for the respective image (for RNA panel)
      environment_cells_RNA = data.frame(matrix(ncol=ncol(input_df_RNA), nrow = 0))
      colnames(environment_cells_RNA) = colnames(input_df_RNA)
  
    # create list for plots/polygon objects
      polygon_plots = ggplot()
      
    # loop through all clusters of k in image j and extract coordinates of enlarged polygon
      for(i in unique(input_df_RNA[input_df_RNA$ImageNumber == j & input_df_RNA$cluster != 0,]$cluster)){
        
        cur_df <- input_df_RNA[input_df_RNA$cluster == i & input_df_RNA$cluster != 0,]
  
        ## Compute concave hull for each cluster
        if (nrow(cur_df) > 2){
            coords <- as.matrix(cbind(cur_df$X, cur_df$Y))
            hull = data.table::as.data.table(concaveman::concaveman(coords, concavity = 1))
            colnames(hull) <- c("X", "Y")
  
            # return common rows between input and the hull
            # first, we need to round the digits of cur_df since concaveman does not return full-length coordinates
            # (see github issue https://github.com/joelgombin/concaveman/issues/13)
            # is obsolete once the bug is fixed
            cur_df_2 <- data.table::copy(cur_df)
            cur_df_2$X <- round(cur_df_2$X, digits = 4)
            cur_df_2$Y <- round(cur_df_2$Y, digits = 4)
  
            # create polygon object
            polygon = sf::st_polygon(list(as.matrix(hull)))
          }
  
        # buffer/enlarge polygon by "distance" pixels
        polygon_buff = sf::st_buffer(polygon, distance)
        polygon_buff_sfc = sf::st_sfc(polygon_buff)
  
        polygon_plots <- polygon_plots +
          geom_sf(data = polygon, fill = NA, size = 4, col = "deepskyblue2") +
          geom_sf(data = polygon_buff_sfc, fill = NA, size=4, col = "blue")
  
        # return coordinates of the enlarged polygon
        enlarged_coordinates = as.data.frame(polygon_buff[1])
        
        # define cells within this enlarged polygon in the matched ROI from the protein panel
        intersect = sf::st_intersects(polygon_buff_sfc, cells_sfc_protein)
        intersect = rownames(as.data.frame(cells_protein[intersect[[1]],]))
        environment_cells = rbind(environment_cells, input_df_protein[input_df_protein$cellID %in% intersect,])
        
        # define cells within this enlarged polygon in the matched ROI from the RNA panel
        intersect_RNA = sf::st_intersects(polygon_buff_sfc, cells_sfc)
        intersect_RNA = rownames(as.data.frame(cells[intersect_RNA[[1]],]))
        environment_cells_RNA = rbind(environment_cells_RNA, input_df_RNA[input_df_RNA$cellID %in% intersect_RNA,])
      }
    
    # Add cells to the data frame for k
    k_env_cells = rbind(k_env_cells, environment_cells)
    k_env_cells_RNA = rbind(k_env_cells_RNA, environment_cells_RNA)
  }

# Add environment column for cytokine k to protein sce object
sce_protein$environment <- ifelse(sce_protein$cellID %in% k_env_cells$cellID, 1, 0)
names(colData(sce_protein))[length(names(colData(sce_protein)))] <- paste0(k, "_environment")

# Add environment column for cytokine k to RNA sce object
sce$environment <- ifelse(sce$cellID %in% k_env_cells_RNA$cellID, 1, 0)
names(colData(sce))[length(names(colData(sce)))] <- paste0(k, "_environment")

}

```

## Save protein SCE object with cytokine environment information

```{r save-protein-sce}
saveRDS(sce_protein, "output/ProteinPanel/sce_ProteinPanel_subtypes_cytokine_environments.rds")
saveRDS(sce, "output/RNApanel/sce_RNApanel_cytokines_communities.rds")
```

## Compare cell type proportions in cytokine environments: RNA vs protein panel

```{r environments-compare-panels}
# Rename and match celltypes for the two panels
sce_RNA_simple <- sce
sce_RNA_simple[,which(sce_RNA_simple$celltype %in% c("stromal_SMA+", "stromal_undefined"))]$celltype <- "stromal"
sce_RNA_simple[,which(sce_RNA_simple$celltype == "Bcell")]$celltype <- "B_cell"
sce_RNA_simple[,which(sce_RNA_simple$celltype %in% c("NK", "T_CD4", "T_CD8"))]$celltype <- "T_NK"
sce_protein_simple <- sce_protein
sce_protein_simple[,which(sce_protein_simple$celltype %in% c("fibroblast", "stromal_undefined", "plasma_cell", "pDC"))]$celltype <- "stromal"

# Get celltype proportions
coldata_RNA <- data.frame(colData(sce_RNA_simple))
coldata_protein <- data.frame(colData(sce_protein_simple))
tab_RNA <- data.frame(matrix(ncol = 8, nrow = 0))
colnames(tab_RNA) <- levels(factor(sce_RNA_simple$celltype))
tab_protein <- tab_RNA
  
for (k in (colnames(coldata_RNA)[(ncol(coldata_RNA)-10):ncol(coldata_RNA)])) {
  k_sub_RNA <- coldata_RNA[coldata_RNA[,k] == 1,]
  tab_RNA[k,] <- prop.table(table(k_sub_RNA$celltype))
  k_sub_protein <- coldata_protein[coldata_RNA[,k] == 1,]
  tab_protein[k,] <- prop.table(table(k_sub_protein$celltype))  
}

tab_RNA$env <- rownames(tab_RNA)
tab_RNA$panel <- "RNApanel"
tab_RNA <- pivot_longer(tab_RNA, 1:8, names_to = "celltype", values_to = "frequency")

tab_protein$env <- rownames(tab_protein)
tab_protein$panel <- "ProteinPanel"
tab_protein <- pivot_longer(tab_protein, 1:8, names_to = "celltype", values_to = "frequency")

tab_comb <- rbind(tab_protein, tab_RNA)

ggplot(tab_comb, aes(panel, y = frequency, fill=celltype)) +
  geom_bar(stat="identity")+
  #facet_wrap(~env, ncol=6)+
  theme(axis.title.x=element_blank(), axis.title.y = element_blank())+
  theme(panel.background = element_blank(), axis.text.x = element_text(angle = 30))

```


## Calculate cell type enrichment in specific environments

```{r cytokine-environment-celltype-enrichment, fig.width=10, fig.height=7}
coldata_protein <- data.frame(colData(sce_protein))

#Calculate enrichment of cell subtypes in environment k
enrichment <- data.frame(matrix(ncol=length(levels(as.factor(sce_protein$subtype))), nrow = 0))
colnames(enrichment) = levels(as.factor(sce_protein$subtype))

for (k in colnames(coldata_protein)[24:33]) {
  sce_sub_k <- coldata_protein[coldata_protein[,k] == 1,]
  enrichment[k,] <- prop.table(table(sce_sub_k$subtype)) / prop.table(table(sce_protein$subtype))
}

# Plot general enrichment of cell subtypes
enrichment$cluster <- rownames(enrichment)
enrichment_long <- pivot_longer(enrichment, 1:20, names_to = "celltype", values_to = "enrichment")
ggplot(enrichment_long, aes(cluster, enrichment, fill = cluster))+
  geom_bar(stat="identity")+
  facet_wrap(~celltype, scales = "free")+
  theme(panel.background = element_blank(),
        axis.text.x = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  geom_hline(yintercept = 1, linetype = "dashed")


## Calculate individually for each image (to prevent bias from high-infiltration images)

enrichment_image <- data.frame(matrix(ncol=4, nrow = 0))

for (k in colnames(coldata_protein)[24:33]) {
  for(j in unique(as.character(coldata_protein$ImageNumber))) {
    sce_sub_k <- coldata_protein[coldata_protein[,k] == 1,]
    sce_sub_kj <- sce_sub_k[sce_sub_k[,"ImageNumber"] == j,]
    prop_kj <- prop.table(table(sce_sub_kj$subtype))
    prop_j <- prop.table(table(sce_protein[,which(sce_protein$ImageNumber == j)]$subtype))[names(prop_kj)]
    enrichment_j <-  prop_kj/prop_j 
    if(length(prop_kj > 0)) {
          j_df <- data.frame(ImageNumber = j,
                      cytokine = k,
                      enrichment = enrichment_j)
    enrichment_image <- rbind(enrichment_image, j_df)
    }
  }
}

colnames(enrichment_image)[3:4] = c("subtype", "enrichment")

ggplot(enrichment_image, aes(cytokine, enrichment, fill = cytokine))+
  geom_boxplot()+
  facet_wrap(~subtype, scales = "free", ncol = 4)+
  theme(panel.background = element_blank(),
        axis.text.x = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  geom_hline(yintercept = 1, linetype = "dashed")

```

### Leaving out single samples

```{r cytokine-environment-celltype-enrichment-leave-one-out, fig.width=10, fig.height=7}
# leave_out <- "TBB184"
# sce_protein_min1 <- sce_protein[,which(sce_protein$sample != leave_out)]
# coldata_protein_min1 <- data.frame(colData(sce_protein_min1))
# 
# 
# #Calculate enrichment of cell subtypes in environment k
# enrichment <- data.frame(matrix(ncol=length(levels(as.factor(sce_protein_min1$subtype))), nrow = 0))
# colnames(enrichment) = levels(as.factor(sce_protein_min1$subtype))
# 
# for (k in colnames(coldata_protein_min1)[24:33]) {
#   sce_sub_k <- coldata_protein_min1[coldata_protein_min1[,k] == 1,]
#   enrichment[k,] <- prop.table(table(sce_sub_k$subtype)) / prop.table(table(sce_protein_min1$subtype))
# }
# 
# # Plot general enrichment of cell subtypes
# enrichment$cluster <- rownames(enrichment)
# enrichment_long <- pivot_longer(enrichment, 1:20, names_to = "celltype", values_to = "enrichment")
# ggplot(enrichment_long, aes(cluster, enrichment, fill = cluster))+
#   geom_bar(stat="identity")+
#   facet_wrap(~celltype, scales = "free")+
#   theme(panel.background = element_blank(),
#         axis.text.x = element_blank(),
#         panel.border = element_rect(color = "black", fill = NA, size = 1),
#         strip.background = element_blank(),
#         legend.background = element_blank())+
#   geom_hline(yintercept = 1, linetype = "dashed")

```

### Enrichment of different T cell subtypes

```{r cytokine-environment-enrichment_TNK, fig.width=8.5, fig.height=6}
#Calculate enrichment of T cell subtypes in environment k, exclude T_AllHigh
sce_TNK <- sce_protein[,which(sce_protein$celltype == "T_NK" & sce_protein$subtype != "T_AllHigh")]
coldata_TNK <- data.frame(colData(sce_TNK))

enrichment_TNK <- data.frame(matrix(ncol=length(levels(as.factor(sce_TNK$subtype))), nrow = 0))
colnames(enrichment_TNK) = levels(as.factor(sce_TNK$subtype))
  
 for (k in colnames(coldata_TNK)[24:33]) { 
  sce_TNK_sub_k <- coldata_TNK[coldata_TNK[,k] == 1,]
  enrichment_TNK[k,] <- prop.table(table(sce_TNK_sub_k$subtype)) / prop.table(table(sce_TNK$subtype))
}

# Plot enrichment of T cell subtypes OUT OF ALL TNK CELLS in all cytokine communities
enrichment_TNK$cluster <- rownames(enrichment_TNK)
enrichment_long_TNK <- pivot_longer(enrichment_TNK, 1:6, names_to = "subtype", values_to = "enrichment")

ggplot(enrichment_long_TNK, aes(cluster, enrichment, fill = cluster))+
  geom_bar(stat="identity")+
  facet_wrap(~subtype, scales = "free", ncol = 4)+
  theme(panel.background = element_blank(),
        axis.text.x = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  geom_hline(yintercept = 1, linetype = "dashed")


## Calculate individually for each sample (to prevent sample-bias)

enrichment_TNK <- data.frame(matrix(ncol=4, nrow = 0))

for (k in colnames(coldata_TNK)[24:33]) {
  for(j in unique(as.character(coldata_TNK$sample))) {
    sce_TNK_sub_k <- coldata_TNK[coldata_TNK[,k] == 1,]
    sce_TNK_sub_kj <- sce_TNK_sub_k[sce_TNK_sub_k[,"sample"] == j,]
    prop_kj <- prop.table(table(sce_TNK_sub_kj$subtype))
    prop_j <- prop.table(table(sce_TNK[,which(sce_TNK$sample == j)]$subtype))[names(prop_kj)]
    enrichment_j <-  prop_kj/prop_j 
    if(length(prop_kj > 0)) {
          j_df <- data.frame(sample = j,
                      cytokine = k,
                      enrichment = enrichment_j)
          j_df$enrichment.Var1 <- as.character(j_df$enrichment.Var1)
          
          for (i in unique(sce_TNK$subtype)) {
            if(!(i %in% j_df$enrichment.Var1)){
              j_df <- rbind(j_df, c(j, k, i, 0))
              }     
          }
          enrichment_TNK <- rbind(enrichment_TNK, j_df)
    }
  }
}

colnames(enrichment_TNK)[3:4] = c("T_subtype", "enrichment")
enrichment_TNK$enrichment <- as.numeric(enrichment_TNK$enrichment)

ggplot(enrichment_TNK, aes(cytokine, enrichment, fill = cytokine))+
  geom_boxplot()+
  facet_wrap(~T_subtype, scales = "free", ncol = 3)+
  theme(panel.background = element_blank(),
        axis.text.x = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  geom_hline(yintercept = 1, linetype = "dashed")


## Calculate individually for each image (to prevent bias from high-infiltration images)

enrichment_image <- data.frame(matrix(ncol=4, nrow = 0))

for (k in colnames(coldata_TNK)[24:33]) {
  for(j in unique(as.character(coldata_TNK$ImageNumber))) {
    sce_sub_k <- coldata_TNK[coldata_TNK[,k] == 1,]
    sce_sub_kj <- sce_sub_k[sce_sub_k[,"ImageNumber"] == j,]
    prop_kj <- prop.table(table(sce_sub_kj$subtype))
    prop_j <- prop.table(table(sce_TNK[,which(sce_TNK$ImageNumber == j)]$subtype))[names(prop_kj)]
    enrichment_j <-  prop_kj/prop_j 
    if(length(prop_kj > 0)) {
          j_df <- data.frame(ImageNumber = j,
                      cytokine = k,
                      enrichment = enrichment_j)
          j_df$enrichment.Var1 <- as.character(j_df$enrichment.Var1)
          
          for (i in unique(sce_TNK$subtype)) {
            if(!(i %in% j_df$enrichment.Var1)){
              j_df <- rbind(j_df, c(j, k, i, 0))
            }
          }
    enrichment_image <- rbind(enrichment_image, j_df)
    }
  }
}

colnames(enrichment_image)[3:4] = c("subtype", "enrichment")
enrichment_image$enrichment <- as.numeric(enrichment_image$enrichment)

ggplot(enrichment_image, aes(cytokine, enrichment, fill = cytokine))+
  geom_boxplot(outlier.shape = NA)+
  coord_cartesian(ylim = c(0, 3))+
  facet_wrap(~subtype, scales = "free", ncol = 3)+
  theme(panel.background = element_blank(),
        axis.text.x = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  geom_hline(yintercept = 1, linetype = "dashed")
  

```
