==> analysis/12_RNApanel_CXCL13_protein.Rmd <==
---
title: "12_RNApanel_CXCL13_protein"
author: "SandraTietscher"
date: "2020-11-11"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction

This script adresses the question of how far away from the CXCL13 transcript-expressing cell CXCL13 protein can still be detected.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message= FALSE)
```


# Load packages, helper functions and data

```{r message=FALSE, warning=FALSE}
library(SingleCellExperiment)
library(ggplot2)
library(scater)
library(igraph)
library(reshape2)
library(cowplot)
library(ggridges)
library(tidyverse)
library(viridis)
library(dplyr)
library(sp)
library(sf)
library(RANN)
library(concaveman)
library(ggpubr)
library(data.table)
library(patchwork)

sce = readRDS(file = "output/RNApanel/sce_RNApanel_cytokines_communities.rds")

TIG2 <- c("TBB111", "TBB129", "TBB165", "TBB171", "TBB184", "TBB338")

# Read object relationship file (output from Cellprofiler)
dat_relation <- fread("data/cpout/Object_relationships_RNApanel.csv")
dat_relation$cellA <- paste0(dat_relation$`First Image Number`, "_", dat_relation$`First Object Number`)
dat_relation$cellB <- paste0(dat_relation$`Second Image Number`, "_", dat_relation$`Second Object Number`)
```

# Compare CXCL13 protein status to CXCL13 mRNA status

```{r-RNA-protein-boxplot, fig.width=3, fig.height=4}
sce_T <- sce[,which(sce$celltype %in% c("T_CD8", "T_CD4"))]
CXCL13_exp <- data.frame(mRNA = as.factor(sce_T$CXCL13_mRNA_expression),
                         protein = assay(sce_T, "counts")["CXCL13_protein",])

ggplot(CXCL13_exp, aes(x=mRNA, y = protein))+
  #geom_violin(width=1, fill = "black")+
  geom_boxplot(outlier.shape = NA)+
  stat_compare_means(label = "p.format", method = "wilcox", label.x.npc = "middle")+
  theme(panel.border = element_rect(color = "black", fill = NA, size = 1),
        panel.background = element_blank())+
  xlab("CXCL13 mRNA expression")+
  ylab("CXCL13 protein counts")+
  coord_cartesian(ylim = c(0,10))
```

# Calculate distance to the next CXCL13_mRNA expressing cell (for each cell)

For images in which there is no CXCL13_mRNA positive cell --> set distance parameter to NA and exclude for further analysis.

```{r calculate-distance}
cur_sce <- data.frame(colData(sce))
cellIDs_cxcl13 <- cur_sce[cur_sce[,"CXCL13_mRNA_expression"] == 1,]
new_sce <- data.frame(matrix(ncol=length(colnames(cur_sce))+1, nrow = 0))

# For each image separately: nearest neighbour search

for (j in unique(cur_sce$ImageNumber)) {
  cur_sce_j <- filter(cur_sce, ImageNumber == j)
  cxcl13_j <- filter(cellIDs_cxcl13, ImageNumber == j)[,c("Center_X", "Center_Y")]
  if (nrow(cxcl13_j) == 0) {
    cur_sce_j$cxcl13_distance <- NA
  } else {
    j_coord <- cur_sce_j[,c("Center_X", "Center_Y")]
    nn <- RANN::nn2(cxcl13_j, j_coord)
    cur_sce_j$cxcl13_distance <- nn$nn.dists[,1]
  }
  new_sce <- rbind(cur_sce_j, new_sce)
}
```

# Plot distance in correlation with CXCL13_protein expression

```{r distance-cxcl13-expression}
new_sce$CXCL13_protein <- assay(sce, "counts")["CXCL13_protein",]
new_sce_sub <- new_sce[!is.na(new_sce$cxcl13_distance),]

# Remove outlier cells (mean counts > 30)
new_sce_sub <- new_sce_sub[new_sce_sub$CXCL13_protein <= 30,]

# Bin the distances
breaks <- seq(from = 0, to = 1000, by = 5)
new_sce_sub$distance_tag <- cut(new_sce_sub$cxcl13_distance, breaks = breaks, include.lowest = TRUE, right = TRUE)

ggplot(new_sce_sub, aes(cxcl13_distance, CXCL13_protein)) +
  geom_point()+
  geom_smooth(method = lm, color = "red", se=FALSE)+
  stat_cor(method="spearman")+
  theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())

```

# Identify direct neighbours of CXCL13_RNA positive cells

```{r neighbour-cxcl13-expression, fig.width=5,fig.height=4}
cxcl13_shortID <- paste0(cellIDs_cxcl13$ImageNumber, "_", cellIDs_cxcl13$CellNumber)

# Subset neighbourhood list
dat_relation_sub <- dat_relation %>% filter(cellA %in% cxcl13_shortID | cellB %in% cxcl13_shortID)
neighbours <- unique(c(dat_relation_sub$cellA, dat_relation_sub$cellB))
neighbours <- setdiff(neighbours, cxcl13_shortID)

# Add information to colData
cur_sce$shortID <- paste0(cur_sce$ImageNumber, "_", cur_sce$CellNumber)
cur_sce$cxcl13_status <- ifelse(cur_sce$shortID %in% cxcl13_shortID, "mRNA_expressing", ifelse(cur_sce$shortID %in% neighbours, "neighbour", "none"))
cur_sce$cxcl13_counts <- assay(sce, "counts")["CXCL13_protein",]

# Plot
ggplot(cur_sce, aes(cxcl13_status, cxcl13_counts))+
         geom_boxplot(outlier.shape = NA)+
    coord_cartesian(ylim=c(0, 8))+
    theme(axis.line.x = element_line(colour = "black", size = 0.25),
        axis.line.y = element_line(colour = "black", size = 0.25),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_rect(colour = "black", fill=NA, size=1),
        panel.background = element_blank())

```

# CXCL13+ T cells: Single cells vs CXCL13-clusters

```{r CXCL13-single-cluster}
sce_T_CXCL13 <- sce[,which(sce$subtype %in% c("T_CD4_CXCL13", "T_CD8_CXCL13"))]
sce_T_CXCL13$CXCL13_cluster <- ifelse(sce_T_CXCL13$cxcl13_mrnaonly_clust == 0, "single", "cluster")
tab_subtype <- as.data.frame(prop.table(table(sce_T_CXCL13$subtype, sce_T_CXCL13$CXCL13_cluster), margin = 1))
tab_TIG <- as.data.frame(prop.table(table(sce_T_CXCL13$TIG, sce_T_CXCL13$CXCL13_cluster), margin = 1))
tab_TLS <- as.data.frame(prop.table(table(sce_T_CXCL13$TLS, sce_T_CXCL13$CXCL13_cluster), margin = 1))

p_subtype <- ggplot(tab_subtype, aes(Var1, Freq, fill = Var2))+
    geom_bar(stat = "identity")+
    scale_fill_manual(values = c("#E69F00", "#006600"))+
    theme(axis.title.x=element_blank(),panel.background = element_blank(), axis.text.x = element_text(angle = 60, hjust = 1), legend.position = "none")

p_TIG <- ggplot(tab_TIG, aes(Var1, Freq, fill = Var2))+
    geom_bar(stat = "identity")+
    scale_fill_manual(values = c("#E69F00", "#006600"))+
    theme(axis.title.x=element_blank(),panel.background = element_blank(), axis.text.x = element_text(angle = 60, hjust = 1), legend.position = "none")

p_TLS <- ggplot(tab_TLS, aes(Var1, Freq, fill = Var2))+
    geom_bar(stat = "identity")+
    scale_fill_manual(values = c("#E69F00", "#006600"))+
    theme(axis.title.x=element_blank(),panel.background = element_blank(), axis.text.x = element_text(angle = 60, hjust = 1), legend.title = element_blank())

p <- (p_subtype + p_TIG) + p_TLS
p

```

```{r CXCL13-single-cluster-sample}
tab_sample <- as.data.frame(prop.table(table(sce_T_CXCL13$sample, sce_T_CXCL13$CXCL13_cluster), margin = 1))
tab_sample <- filter(tab_sample, Var2 == "cluster")
colnames(tab_sample) <- c("sample", "cluster", "proportion")
tab_sample$TIG <- ifelse(tab_sample$sample %in% TIG2, "TIG2", "TIG3")

ggplot(tab_sample, aes(x= TIG, y = proportion, color=TIG))+
  geom_boxplot()+
  geom_point()+
  theme(axis.ticks.x = element_blank(),
        axis.text.x=element_text(angle = 90, hjust=1),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_rect(color = "black", fill = NA, size = 1),
        strip.background = element_blank(),
        legend.background = element_blank())+
  ylab("Of total [%]")+
  stat_compare_means(method = "wilcox.test", aes(label=..p.format..), label.x.npc = 0.5)+
  ggtitle("Proportion of cells that are part of a CXCL13+ cluster")

```
